% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{ale2violin}
\alias{ale2violin}
\title{Prepare ALE data for violin plots}
\usage{
ale2violin(
  iMatrixAle = NULL,
  iMatrixAleGrp = NULL,
  groups,
  facet_groups = groups,
  facet_name = "Group",
  maxGroupMeanALE = 2,
  removeAboveAleNum = "ale3",
  maxGroupMeanFloor = 0,
  returnAll = FALSE,
  geneLists,
  lineAlpha = 0.1,
  subsetFunc = NULL,
  make_ggplots = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{groups}{vector of group labels, named by \code{colnames(iMatrixALE)}.}

\item{facet_groups}{vector of group labels, named by \code{colnames(iMatrixALE)},
as a possible alternative to using the \code{groups}, for example
for higher level grouping.}

\item{facet_name}{character string used to label the \code{facet_groups}.}

\item{maxGroupMeanALE}{numeric value indicating the threshold for
including an ALE in the output data, where the max group mean
(the highest group mean) is at least this value.}

\item{removeAboveAleNum}{character string of the ALE colname to remove,
restricting data to include only ALE values below this number. For
example "ale3" would remove "ale3" and all higher ALE numbers,
thereby restricting data to "ale1" and "ale2".}

\item{maxGroupMeanFloor}{numeric threshold used as a noise floor.}

\item{returnAll}{logical indicating whether to return intermediate
data formats in the output results list.}

\item{geneLists}{list containing vectors of genes, or a data.frame
with two columns "gene_name" and "geneList", used
to assign genes to one or more lists in the resulting violin plot.}

\item{lineAlpha}{numeric value of alpha transparency, scaled between
0 and 1, used to draw lines from "_ale1" to "_ale2" on the
violin plot.}

\item{subsetFunc}{optional function that takes the tall format data used in the
primary violin plot, and returns data in the same format after
applying logic specific for filtering this data. Intended to
restrict display of genes to \code{groups} or \code{facet_groups} that are
relevant to each \code{geneLists} entry.}

\item{make_ggplots}{logical indicating whether to create plot objects
using ggplot2.}

\item{verbose}{logical indicating whether to print verbose output}

\item{...}{additional arguments are ignored.}

\item{iMatrixALE}{numeric matrix of expression data containing
ALE rows, as output from \code{tx2ale()}. Each rowname is expected
to have a suffix "_ale1" where the number represents the stranded
order of ALE elements in a given gene. Therefore "_ale1" is the
shortest form, closest to the 5-prime end of the transcript, and
"_ale2" is the next ALE element downstream, and so on.}
}
\description{
Prepare ALE data for violin plots
}
\details{
This function takes output from \code{tx2ale()}, applies some filtering
to the output data, then returns a tall data.frame sufficient
for viewing as a violin plot using \code{ggplot2::geom_violin()}.
}
\seealso{
Other ALE and codon functions:
\code{\link[=codonUsage2df]{codonUsage2df()}},
\code{\link[=dna2codon]{dna2codon()}},
\code{\link[=getFirstStrandedFromGRL]{getFirstStrandedFromGRL()}},
\code{\link[=jamCai]{jamCai()}},
\code{\link[=tx2ale]{tx2ale()}}
}
\concept{ALE and codon functions}
