% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gr-coverage.R
\name{combineGRcoverage}
\alias{combineGRcoverage}
\title{Combine GRanges coverage replicates}
\usage{
combineGRcoverage(
  gr,
  covNames = NULL,
  covName = NULL,
  strands = NULL,
  scaleFactors = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{GRanges} object containing coverage data in columns
containing NumericList class data.}

\item{covNames}{\code{character} vector of \code{colnames(GenomicRanges::values(gr))}
representing columns in \code{gr} that contain coverage data
in NumericList format, for example data prepared
with \code{getGRcoverageFromBw()}.}

\item{covName}{\code{character} vector with length equal to
\code{length(covNames)} representing the \code{sample_id} for each
\code{covNames} entry.}

\item{strands}{\code{character} vector, or NULL, indicating the strand
for which the coverage data was obtained. When \code{NULL} the strand
is inferred by the presence of any negative values.}

\item{scaleFactors}{\code{numeric} vector length equal to \code{length(covNames)}
or expanded to that length. Values are multiplied by each
coverage data result, intended to apply a normalization
to each coverage value. A \code{-1} value can also be used
to flip the score of negative strand data, in the event the
source coverage data is scored only using positive values.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
GRanges object whose colnames contain the
\code{covName} (\code{sample_id}) for each observed strand,
with coverage combined taking the sum of individual
coverages after multiplying each by \code{scaleFactors}.
}
\description{
Combine GRanges coverage replicates
}
\details{
This function takes a GRanges object as output from
\code{getGRcoverageFromBw()} and combines the coverages into
one coverage per strand for each \code{covName} (equivalent
to \code{sample_id}). Each coverage value is multiplied by
its \code{scaleFactors} value, then the sum is returned for
each strand, for each \code{covName} (\code{sample_id}).

The strand is inferred by the
presence of negative values, where any negative value
indicates the column is negative strand.
}
\seealso{
Other Internal utility functions:
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
