% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{compressGRgaps}
\alias{compressGRgaps}
\title{Compress GRanges gaps to a fixed width}
\usage{
compressGRgaps(GR, gapWidth = NULL, keepValues = FALSE,
  upstream = 10000, upstreamGapWidth = gapWidth * 3,
  downstream = 10000, downstreamGapWidth = gapWidth * 3,
  refGR = NULL, refGRnames = NULL, verbose = FALSE, ...)
}
\description{
Compress GRanges gaps to a fixed width
}
