% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{defineDetectedTx}
\alias{defineDetectedTx}
\title{Define detected transcripts}
\usage{
defineDetectedTx(
  iMatrixTx = NULL,
  iMatrixTxGrp = NULL,
  iMatrixTxTPM = NULL,
  iMatrixTxTPMGrp = NULL,
  groups = NULL,
  tx2geneDF = NULL,
  cutoffTxPctMax = 10,
  cutoffTxExpr = 5,
  cutoffTxTPMExpr = 0.1,
  txColname = "transcript_id",
  geneColname = "gene_name",
  zeroAsNA = TRUE,
  applyTxPctTo = c("TPM", "counts", "both", "either"),
  useMedian = FALSE,
  floorTPM = 0.001,
  floorCounts = 0.001,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{iMatrixTx}{numeric matrix of read counts (or pseudocounts) with
transcript rows and sample columns. This data is assumed to be
log2-transformed, and if any value is higher than 50, it will
be log2-transformed with \code{log2(x+1)}.}

\item{iMatrixTxGrp}{numeric matrix of read counts averaged by sample
group. If this matrix is not provided, it will be calculated
from \code{iMatrixTx}
using \code{jamba::rowGroupMeans()} and the \code{groups} parameter.
This data is assumed to be
log2-transformed, and if any value is higher than 50, it will
be log2-transformed with \code{log2(x+1)}.}

\item{iMatrixTxTPM}{numeric matrix of TPM values, with sample columns
and transcript rows. Note that if this parameter is not supplied,
the counts in \code{iMatrixTx} will be used to determine the percent
max isoform expression.}

\item{iMatrixTxTPMGrp}{numeric matrix of TPM values averaged by sample
group. If this matrix is not provided, it will be calculated
from \code{iMatrixTxTPM}
using \code{jamba::rowGroupMeans()} and the \code{groups} parameter.}

\item{groups}{vector of group labels, either as character vector
or factor. It should be named by \code{colnames(iMatrixTx)}.}

\item{tx2geneDF}{data.frame with colnames including
\code{c("transcript_id","gene_name")}, where the values in the
\code{"transcript_id"} column must match the \code{rownames(iMatrixTx)}.}

\item{cutoffTxPctMax}{numeric value scaled from 0 to 100
indicating the percentage of
the maximum isoform expression per gene, for an alternate isoform
to be considered for detection.}

\item{cutoffTxExpr}{numeric value indicating the minimum group mean
counts in \code{iMatrixTxGrp} for a transcript to be considered for
detection.}

\item{cutoffTxTPMExpr}{numeric value indicating the minimum group mean
TPM in \code{iMatrixTxTPMGrp} for a transcript to be considered for
detection.}

\item{txColname, geneColname}{the \code{colnames(tx2geneDF)} representing
the \code{rownames(iMatrixTx)} matched by \code{tx2geneDF[,txColname]},
and the associated genes given by \code{tx2geneDF[,geneColname]}.
Note that \code{detectedTx} must also contain values in
\code{rownames(iMatrixTx)} and \code{tx2geneDF[,txColname]}.}

\item{zeroAsNA}{logical indicating whether values of zero
(or less than zero) should
be treated as \code{NA} values, thus removing them from mean
calculations. This argument is only relevant when \code{iMatrixTxGrp}
and \code{iMatrixTxTPMGrp} are not supplied.
Argument \code{zeroAsNA=TRUE} is recommended when using kmer
quantitation tools such as Salmon or Kallisto, which can
sometimes allocate all expression to one or another transcript
isoform when two isoforms are nearly identical. Also use
\code{TRUE} when a value of zero represents the absense of data.}

\item{applyTxPctTo}{\code{character} string indicating how to apply the
\code{cutoffTcPctMax} threshold. This argument is only used when
both \code{iMatrixTx} and \code{iMatrixTxTPM} are supplied, as follows:
\code{"TPM"}: uses only \code{iMatrixTxTPM} data; \code{"counts"} uses only
\code{iMatrixTx} data; \code{"both"} requires both \code{iMatrixTx} and
\code{iMatrixTxTPM} data meet the threshold; \code{"either"} requires
that one or both of \code{iMatrixTx} and \code{iMatrixTxTPM} meet the
threshold.}

\item{useMedian}{logical indicating whether to use group median
values instead of group mean values.}

\item{floorTPM, floorCounts}{\code{numeric} value indicating the floor
value to use when isoform TPM or counts are below \code{1}, used
to prevent divide-by-zero when all isoforms are \code{0}.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\value{
List with the following elements:
\describe{
\item{txExprGrpTx}{Numeric matrix representing the expression
counts per transcript, grouped by \code{"gene_name"}.}
\item{txPctMaxTxGrpAll}{Numeric matrix representing the
percent expression of each transcript isoform per gene, as
compared to the highest expression of isoforms for that gene,
using \code{iMatrixTxGrp} data.
(New to verion 0.0.61.900.)}
\item{txPctMaxTxTPMGrpAll}{Numeric matrix representing the
percent expression of each transcript isoform per gene, as
compared to the highest expression of isoforms for that gene,
using \code{iMatrixTxTPMGrp} data. This data is returned only
if \code{iMatrixTxTPM} or \code{iMatrixTxTPMGrp} were supplied.
(New to verion 0.0.61.900.)}
\item{txPctMaxGrpAll}{Numeric matrix representing the
percent max expression used for filtering, after applying
\code{applyTxPctTo}: \code{"counts"} uses \code{txPctMaxTxGrpAll};
\code{"TPM"} uses \code{txPctMaxTxTPMGrpAll}; \code{"both"} uses the higher
of \code{txPctMaxTxGrpAll} and \code{txPctMaxTxTPMGrpAll}; \code{"either"}
uses the lower of \code{txPctMaxTxGrpAll} and \code{txPctMaxTxTPMGrpAll}.}
\item{txExprGrpAll}{Numeric matrix of sample group counts,
exponentiated and rounded to integer values.}
\item{txTPMExprGrpAll}{Numeric matrix of sample group TPM values,
exponentiated and rounded to integer values.}
\item{txFilterM}{Numeric matrix indicating whether each isoform met
the criteria to be considered detected. The criteria must be
met in the same group for an isoform to be considered detected.}
\item{detectedTx}{Character vector of transcripts, as defined by
the \code{rownames(iMatrixTx)}.}
}
}
\description{
Define detected transcripts
}
\details{
This function aims to combine evidence from RNA-seq sequence
read counts (or pseudocounts from a kmer tool such as Salmon or
Kallisto), along with alternative TPM quantitation, to determine
the observed "detected" transcript space for a given experiment.

Each input data matrix is assumed to be appropriately log-transformed,
typically using \code{log2(1+x)}. If any value is \verb{>= 50} then the data
matrix will be log2-transformed using \code{log2(1+x)}.

The criteria must be met in at least one sample group, but all
criteria must be met in the same sample group for an isoform to
be considered "detected".

In our experience the use of TPM values appears more robust and
is conceptually the best approach for comparing the relative
quantity of one transcript isoform to another. Our reasoning is
that TPM is intended to be roughly a molar quantity of transcript
molecules, independent of the transcript length, and the potential
for overlapping regions between isoforms. We also recommend the use
of a kmer quantitation method, such as Salmon or Kallisto, which
estimates isoform abundances not by specific read counts, but by
quantifying kmers unique to particular isoforms for a given gene.

In all cases, the thresholds for detection can be modified, however
from our experiences thus far the default values perform reasonably
well at identifying expressed isoforms, while filtering out isoforms
that we considered to be spuriously expressed.

There are three default requirements for a transcript to be considered
"detected".
\enumerate{
\item An isoform must be expressed at least 10\% of the max isoform for
a given gene, using TPM values.
\item An isoform must have at least log2(32) pseudocounts to be considered
detected, based upon our view of Salmon pseudocount data using MA-plots.
\item An isoform must have at least log2(2) TPM units to be considered
detected, based upon our view of Salmon TPM values using MA-plots.
}

Each experiment is likely to be different in terms of total
sequenced reads, quality of read alignment or quantitation to the
transcriptome, etc. We suggest observing MA-plots for the counts and
TPM values, for the point at which the signal substantially increases
from baseline zero. We also plotted the TPM versus count per sample,
noted the point at which the two signals began to correlate. These
observations along with careful review of numerous gene model
transcript isoforms supported our selection of these criteria.

Lastly, the requirement for 10 percent of max isoform expression
was motivated by observing highly expressed genes, which sometimes had
alternative isoforms with extremely low abundance compared to the
most abundant isoform, but which was notably higher than the minimum
for detection. For example Gapdh expression above 100,000 pseudocounts,
may have an isoform with 120 pseudocounts. When we reviewed the sequence
coverage, we could find no compelling evidence to support the minor
isoform, and theorized that the pseudocounts arose from the stochastic
nature of rebalancing relative expression among isoforms.

Note the argument \code{zeroAsNA=TRUE}, which by default treats any
expression value of zero (or less than zero) as \code{NA},
thus removing them from group
mean calculations. When \code{iMatrixTxGrp} and \code{iMatrixTxTPMGrp}
are not supplied, this option is helpful in calculating a more
appropriate group mean expression value, notably when a
value of zero represents absence of data. Any group mean that is
\code{NA} as a result is converted to zero for the purpose of applying
filters.
}
\seealso{
Other Detected transcripts:
\code{\link[=detectedTxInfo]{detectedTxInfo()}}
}
\concept{Detected transcripts}
