% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{df2colorSub}
\alias{df2colorSub}
\title{Convert data.frame to categorical colors}
\usage{
df2colorSub(x, colorSub = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{data.frame}

\item{colorSub}{vector of colors, whose names are intended to match
values in the input data.frame \code{x}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters are passed to \code{colorjam::group2color()}.}
}
\value{
data.frame of the same dimensions as the input \code{x} data.frame,
where values have been substituted with R colors.
}
\description{
Convert data.frame to categorical colors
}
\details{
This function is a temporary placeholder function which simply
provides categorical colors for values in each column of the input
\code{data.frame}. If colors are provided using the named color vector
\code{colorSub}, they are used to substitute values in each column,
otherwise \code{colorjam::group2colors()} is used to create categorical
colors, which itself uses \code{colorjam::rainbowJam()}.

Note that providing \code{colorSub} helps keep colors consistent, otherwise
each column is independently colorized.
}
\examples{
colorSub1 <- colorjam::group2colors(LETTERS[1:6]);
df <- data.frame(one=LETTERS[1:6],
   two=rep(LETTERS[3:4], each=3),
   three=rep(LETTERS[5:6], 3));
dfColors <- df2colorSub(df, colorSub1);
jamba::imageByColors(dfColors, cellnote=df);

# Or in one step
jamba::imageByColors(df2colorSub(df), cellnote=df);

}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
