% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{dfWide2segments}
\alias{dfWide2segments}
\title{Convert data.frame to plotly line segment format}
\usage{
dfWide2segments(x, axes1, axes2, ...)
}
\arguments{
\item{x}{numeric matrix of coordinates, where x,y,z coordinates are
present twice per row, and whose colnames are defined by
\code{axes1} and \code{axes2}.}

\item{axes1}{vector of colnames(x) corresponding to the x,y,z coordinates
of the first point in a line segment. It can be of any length, but
is intended for length 2 or 3.}

\item{axes2}{vector of colnames(x) corresponding to the x,y,z coordinates
of the second point in a line segment. It must be the same length as
\code{axes1}.}

\item{...}{additional parameters are ignored.}
}
\value{
data.frame where every two rows represents a line segment
to be drawn by plotly, followed by a blank line indicating the line
ends.
}
\description{
Convert data.frame to plotly line segment format
}
\details{
Purpose is to take a data.frame containing (x,y) or (x,y,z) coordinates
for two points defined on one row, and create a data.frame with 3 rows
where blank coordinates are used to separate each line segment.
The main benefit in using this function is that it enables vectorized
line segment drawing when used with a plotting function that recognizes
line breaks via empty rows, for example plotly.
}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
