% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{dna2codon}
\alias{dna2codon}
\title{Convert DNA to 3-base codons}
\usage{
dna2codon(x, ...)
}
\arguments{
\item{x}{character vector assumed to contain DNA or RNA nucleotides}

\item{...}{additional arguments are ignored.}
}
\value{
character vector where each element contains three characters,
e.g. \code{all(nchar(dna2codon(x)) == 3)}.
}
\description{
Convert DNA to 3-base codons
}
\details{
This function takes character string input, either as one large
string or vector of strings such as those read from a FASTA file,
and returns a vector of 3-base codons. The final codon must contain
all three character values otherwise it is removed.
}
\examples{
dna <- c("atgggattatag");
dna2codon(dna);

# example adding one extra base, which does not make a full codon
dnaext1 <- c("atgggattataga");
dna2codon(dnaext1);

}
\seealso{
Other ALE and codon functions:
\code{\link[=ale2violin]{ale2violin()}},
\code{\link[=codonUsage2df]{codonUsage2df()}},
\code{\link[=getFirstStrandedFromGRL]{getFirstStrandedFromGRL()}},
\code{\link[=jamCai]{jamCai()}},
\code{\link[=tx2ale]{tx2ale()}}
}
\concept{ALE and codon functions}
