% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-design.R
\name{escapeWhitespaceRegexp}
\alias{escapeWhitespaceRegexp}
\title{Escape whitespace in regular expression patterns}
\usage{
escapeWhitespaceRegexp(x, whitespace = "_ ", maxN = 20, ...)
}
\arguments{
\item{x}{character vector containing a regular expression pattern,
or a character value that should be converted to a regular expression
pattern.}

\item{whitespace}{character vector \code{length=1} containing whitespace
characters, for example \code{" _"} will define space \code{" "} and
underscore \code{"_"} both as whitespace characters.}

\item{maxN}{integer value for maximum iterations of the substitution,
unfortunately the regular expression logic only matches once per
iteration per string.}

\item{...}{additional arguments are ignored.}
}
\description{
Escape whitespace in regular expression patterns
}
\details{
This function is intended to test for unescaped whitespace characters
in a regular expression pattern match string, and replace them with
escaped whitespace characters, possibly expanding the allowed
whitespace characters in the process.
}
\examples{
x <- c("one two three", "one[ ]two three", "one[ 12] two three");
escapeWhitespaceRegexp(x);
# side-by-side summary of input and output
data.frame(input=paste0('"', x, '"'),
   output=paste0('"', escapeWhitespaceRegexp(x), '"'))

}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
