% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{expandExonsGR}
\alias{expandExonsGR}
\title{Expand exon GRanges upstream and downstream}
\usage{
expandExonsGR(inputExonsGR, expandUpDown = c(0, 0),
  expandLabels = c("upstream", "downstream"),
  inputExonsGRlabel = "exonLabel", ...)
}
\description{
Expand exon GRanges upstream and downstream
}
\details{
This function expands an exon to make it wider before and
after the start,end coordinates.
}
