% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{factor2label}
\alias{factor2label}
\title{Convert factor to a factor label}
\usage{
factor2label(
  x,
  valuesL = NULL,
  types = c("count", "none"),
  aggFun = mean,
  digits = 3,
  big.mark = ",",
  itemName = "items",
  ...
)
}
\arguments{
\item{x}{factor vector}

\item{valuesL}{optional list of numeric values, where each vector
has the same length as \code{x}. If it is not a factor, it is
converted to factor, using \code{jamba::mixedSort()} to order
the factor levels.}

\item{types}{character value indicating the summary to use for
the input factor \code{x}.}

\item{aggFun}{summary function used for each vector of numeric values
in \code{valuesL} when supplied.}

\item{digits, big.mark}{arguments passed to \code{base::format()} to
create a suitable text label.}

\item{itemName}{character vector indicating the name to associate to
counts.}

\item{...}{additional arguments are ignored.}
}
\value{
factor vector with the same length as input \code{x} but
where the levels also include summary information, such
as the count of each factor level.
}
\description{
Convert factor to a factor label
}
\details{
This function is intended to take a vector of factor levels
and convert to labels using summary statistics. For example
by default it will add the number of items for each factor
level.

This function is intended to help create useful ordered
factor labels that can be used in a ggplot2 visualization.
}
\examples{
x <- factor(rep(letters[1:5], c(2,4,3,2,1)));
levels(factor2label(x));
factor2label(x);

}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
