% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{flattenExonsByGene}
\alias{flattenExonsByGene}
\title{Flatten transcript exons by gene}
\usage{
flattenExonsByGene(exonsByTx, tx2geneDF, detectedTx, genes = NULL,
  txColname = "transcript_id", geneColname = "gene_name",
  cdsByTx = NULL, cdsByGene = NULL, verbose = FALSE, ...)
}
\arguments{
\item{exonsByTx}{GRangesList named by transcript, containing one or
more GRanges representing exons.}

\item{tx2geneDF}{data.frame containing at least two columns with
transcript and gene annotation, whose colnames are defined by
arguments \code{txColname} and \code{geneColname} respectively.}

\item{detectedTx}{character vector of detected transcripts, used to
subset the overall transcripts prior to producing a flattened gene
exon model.}
}
\value{
GRangesList named by gene, containing non-overlapping GRanges
with exon names as described above.
}
\description{
Flatten transcript exons by gene
}
\details{
This function takes \code{exonsByTx} which is a \code{GRangesList} object
of transcript exons, named by the \code{transcript_id}. It groups transcripts
together by gene, produces a flattened unique set of disjoint exons
which do not overlap. Finally, it labels each exon using a defined
naming scheme:
\itemize{
\item Each contiguous exon is numbered in order, starting at \code{1} for the
first stranded exon for the gene, for example \code{exon1}, \code{exon2},
\code{exon3}.
\item Adjacent exons (within a contiguous exon and no gap between them)
are additionally labeled with a letter suffix to indicate the order
within that exon, for example \code{exon1a}, \code{exon1b}, \code{exon1c}.
}

This function is also useful for generating gene exon models using
only detected transcripts, which can be very helpful in simplifying
gene models composed of a large number of predicted alternative
transcripts. If the method for defining detected transcript isoforms
is relatively effective, the resulting gene models are typically
much improved and easier to interpret visually.
}
\seealso{
Other jam RNA-seq functions: \code{\link{ale2violin}},
  \code{\link{assignGRLexonNames}},
  \code{\link{codonUsage2df}},
  \code{\link{combineGRcoverage}},
  \code{\link{compressPolygonM}},
  \code{\link{defineDetectedTx}},
  \code{\link{detectedTxInfo}}, \code{\link{dna2codon}},
  \code{\link{groups2contrasts}}, \code{\link{jamCai}},
  \code{\link{makeTx2geneFromGtf}},
  \code{\link{make_ref2compressed}},
  \code{\link{runDiffSplice}}, \code{\link{sortSamples}},
  \code{\link{tx2ale}}
}
\concept{GRanges functions}
\concept{jam RNA-seq functions}
