% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getFirstStrandedFromGRL}
\alias{getFirstStrandedFromGRL}
\title{Get first stranded GRanges feature per GRangesList}
\usage{
getFirstStrandedFromGRL(
  grl,
  method = c("direct", "endoapply", "flank"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{grl}{\code{GRangesList}}

\item{method}{\code{character} with method to use, default 'direct':
\itemize{
\item 'direct' uses \code{IRanges::heads()} or
\code{IRanges::tails()} depending upon the strandedness.
This approach is vectorized and is preferred.
\item 'endoapply' uses \code{S4Vectors::endoapply()} to
iterate each GRangesList element, then returns the result
from \code{head()} or \code{tail()}.
This approach is not vectorized and may be substantially
slower than 'direct'.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{GRangesList} containing only the first stranded
\code{GRanges} feature per input \code{GRangesList} element. When
\code{method="flank"} the output contains no metadata values,
but this method is deprecated.
}
\description{
Get first stranded GRanges feature per GRangesList
}
\details{
This function returns the first feature per GRangesList,
relative to the strand of the GRangesList. It assumes each
GRangesList element has only one strand and one seqname,
and will stop otherwise.
}
\examples{
gr12 <- GenomicRanges::GRanges(seqnames=rep("chr1", 9),
   ranges=IRanges::IRanges(
      start=c(100, 200, 400, 500, 300, 100, 200, 400, 600),
      width=c(50,50,50, 50,50,50, 50,50,50)
   ),
   strand=rep(c("+", "-", "+"), c(3,3,3)),
   gene_name=rep(c("GeneA", "GeneB", "GeneC"), each=3)
)
grl <- GenomicRanges::split(gr12, GenomicRanges::values(gr12)$gene_name)
getFirstStrandedFromGRL(grl)

}
\seealso{
Other ALE and codon functions:
\code{\link[=ale2violin]{ale2violin()}},
\code{\link[=codonUsage2df]{codonUsage2df()}},
\code{\link[=dna2codon]{dna2codon()}},
\code{\link[=jamCai]{jamCai()}},
\code{\link[=tx2ale]{tx2ale()}}
}
\concept{ALE and codon functions}
