% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getGRgaps}
\alias{getGRgaps}
\title{Get gaps in GRanges}
\usage{
getGRgaps(gr, strandSpecific = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{gr}{\code{GRanges} object}

\item{strandSpecific}{\code{logical}, default TRUE, whether to determine gaps
within strand.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{getGRLgaps()}.}
}
\description{
Get gaps in GRanges
}
\details{
This function returns the gaps between GRanges regions, calculated
for each chromosome (using \code{GenomicRanges::seqnames(gr)}), and when \code{strandSpecific=TRUE}
it determines gaps in stranded fashion.
}
\seealso{
Other GenomicRanges functions:
\code{\link[=addGRLgaps]{addGRLgaps()}},
\code{\link[=addGRgaps]{addGRgaps()}},
\code{\link[=annotateGRLfromGRL]{annotateGRLfromGRL()}},
\code{\link[=annotateGRfromGR]{annotateGRfromGR()}},
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=closestExonToJunctions]{closestExonToJunctions()}},
\code{\link[=findOverlapsGRL]{findOverlapsGRL()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getGRLgaps]{getGRLgaps()}},
\code{\link[=jam_isDisjoint]{jam_isDisjoint()}},
\code{\link[=sortGRL]{sortGRL()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}}
}
\concept{GenomicRanges functions}
