% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getGRLgaps}
\alias{getGRLgaps}
\title{Get gaps in GRangesList objects}
\usage{
getGRLgaps(grl, strandSpecific = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{grl}{\code{GRangesList} object.}

\item{strandSpecific}{\code{logical}, default TRUE, whether to determine gaps
within strand.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
GRangesList object with gaps for each chromosome and strand
present in each GRanges entry. It does not return gap sequence
at the edges of GRanges regions to the chromosome ends.
}
\description{
Get gaps in GRangesList objects
}
\details{
This function returns gaps between GRanges regions in a GRangesList
object. When \code{strandSpecific=TRUE} is determines gaps per strand,
otherwise strands are converted to \code{"*"}. It will also determine
gaps within chromosome for each GRanges entry in GRangesList.
}
\seealso{
Other GenomicRanges functions:
\code{\link[=addGRLgaps]{addGRLgaps()}},
\code{\link[=addGRgaps]{addGRgaps()}},
\code{\link[=annotateGRLfromGRL]{annotateGRLfromGRL()}},
\code{\link[=annotateGRfromGR]{annotateGRfromGR()}},
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=closestExonToJunctions]{closestExonToJunctions()}},
\code{\link[=findOverlapsGRL]{findOverlapsGRL()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getGRgaps]{getGRgaps()}},
\code{\link[=jam_isDisjoint]{jam_isDisjoint()}},
\code{\link[=sortGRL]{sortGRL()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}}
}
\concept{GenomicRanges functions}
