% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{import_juncs_from_bed}
\alias{import_juncs_from_bed}
\title{Import splice junction data from BED or SJ.out.tab file}
\usage{
import_juncs_from_bed(
  iBed,
  juncNames,
  sample_id,
  scale_factor = 1,
  use_memoise = FALSE,
  memoise_junction_path = "junctions_memoise",
  gr = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{iBed}{\code{character} path or URL to one BED, bigBed, or
'SJ.out.tab' file as produced by STAR during alignment.
\itemize{
\item For 'SJ.out.tab' input, the countUnique value (column 7)
is used, with no other filtering applied.
\item For BED and bigBed input, the score is preferred in the
name column (column 3), primarily because UCSC
tracks limit the score to maximum of 1000.
\item BED name can be numeric, or can be in the form
\code{'JUNC000000_1234'} with 'JUNC' prefix followed by
integer digits, underscore, the integer values.
In this example, the score is \code{'1234'}.
When not all names can be converted to numeric values,
the score column is used.
\item If you see scores with maximum value "1000" the 'score'
column could be the problem.
\item bigBed files are imported using the 'cpp11bigwig' package
which also uses the appropriate index.
}}

\item{juncNames}{\code{character} the name of the junction source file}

\item{sample_id}{\code{character} string representing the sample
identifier.}

\item{scale_factor}{\code{numeric} default 1, adjusts the raw score,
applied by multiplying the scale_factor by each score.}

\item{use_memoise}{\code{logical} default FALSE, whether to cache data
using memoise.}

\item{memoise_junction_path}{\code{character} default 'junctions_memoise'
with default subdirectory for memoise cache files, used only
when \code{use_memoise=TRUE}.}

\item{gr}{\code{GRanges} representing the overall range for which
junction data will be retrieved. Note that any junctions
that span this range, but do not start or end inside this
range, will be removed.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Import splice junction data from BED or SJ.out.tab file
}
\details{
This function is intended to be called internally by
\code{prepareSashimi()}, and is provided primarily to enable
use of \code{memoise::memoise()} to cache results.

This function was refactored in version 0.0.69.900 to
handle either BED format, or \code{"SJ.out.tab"} junction
format as produced by STAR alignment. The method uses
\code{data.table::fread()} then if there are 9 columns,
it assumes the format is \code{"SJ.out.tab"}. Otherwise it
coerces the \code{data.frame} with \code{as(bed, "GRanges")}.

The BED or \code{"SJ.out.tab"} file can be gzipped, provided
\code{data.table()} is able to recognieze and import the
compression format.

Note that bigBed format still cannot be used since the
rtracklayer package does not support that format.

Also note that junctions with \code{score=0} are dropped at
this step, to prevent propagation of junctions with
zero counts.
}
\seealso{
Other Data import functions:
\code{\link[=getGRcoverageFromBw]{getGRcoverageFromBw()}},
\code{\link[=psl2df]{psl2df()}}
}
\concept{Data import functions}
