% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{intercalate}
\alias{intercalate}
\title{Intercalate list values into a vector}
\usage{
intercalate(...)
}
\arguments{
\item{...}{parameters are expected as multiple vectors.}
}
\value{
vector of values derived from each input list \code{...}, with
one element per list in order. In the event any list is shorter than
the others, its values are recycled so each list is equal length.
}
\description{
Intercalate list values into a vector
}
\details{
This function takes a list of vectors, and intercalates them
by inserting alternating values from each vector to make one
combined vector. It is analogous to shuffling a deck of cards,
except that there could be more than two stacks of cards shuffled
together.
}
\examples{
A <- LETTERS[1:10];
B <- letters[1:10];
intercalate(A, B)

}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
