% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{jitter_norm}
\alias{jitter_norm}
\title{Apply jitter using normal distribution}
\usage{
jitter_norm(x, factor = 1, amount = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{factor}{numeric value to define the magnitude of jitter,
multiplied by the default jitter which is 1/10 the median
difference between unique values in \code{x}.}

\item{amount}{optional numeric value indicating a fixed \code{sd}
standard deviation passed to \code{stats::rnorm()}.}

\item{...}{additional arguments are ignored.}
}
\description{
Apply jitter using normal distribution
}
\details{
This function applies a jitter (noise) to points by adding
random values from a normal distribution, where the argument \code{sd}
is used to apply the jitter magnitude.

The default jitter is defined as 1/10 the median difference between
unique, finite input values, which can be scaled using the argument
\code{factor}. Note the use of "unique" input values, which ensures the
presence of duplicate values does not skew the jitter toward zero.
When applied to three dimensions, this results in jitter consistently
scaled relative to the range of values in each dimension. That is,
points will appear to have a radial jitter.
}
\seealso{
Other Plot utility functions:
\code{\link[=bgaPlotly3d]{bgaPlotly3d()}},
\code{\link[=spline3d]{spline3d()}}
}
\concept{Plot utility functions}
