% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny.R
\name{launchSashimiApp}
\alias{launchSashimiApp}
\title{Launch Sashimi R-shiny application}
\usage{
launchSashimiApp(
  ...,
  envir = parent.frame(),
  options = list(width = 1200),
  verbose = FALSE
)
}
\arguments{
\item{...}{additional arguments are passed to \code{sashimiAppConstants()}.}

\item{envir}{\code{environment} default \code{parent.frame()} uses the environment
which called this function.
The environment contains data needed for sashimi plots.
If \code{envir=NULL} by default it will use \code{parent.frame()}.
Otherwise  call \code{sashimiDataConstants()} or \code{sashimiAppConstants()}
which returns an \code{environment} with the necessary data objects,
and that can be passed to this function.}

\item{options}{\code{list} of R-shiny app options, for example two
common options are: \code{host} to indicate the host or IP address
the R-shiny app will bind to respond to requests; and
\code{port} for the port number. For example:
\code{launchSashimiApp(options=list(host="0.0.0.0", port=8080))}
where \code{port="0.0.0.0"} will listen to any request sent
to the current machine, whether by host name, or any valid
IP address; and \code{port=8080} will only listen to port 8080.}
}
\value{
\code{shiny::shiny.appobj} which is a Shiny App object,
suitable to run an R-shiny app by printing to console.

output from \code{shiny::shinyApp()} which is an object of
class "shiny.appobj", whose default print method is to run
the app.
}
\description{
Launch Sashimi R-shiny application
}
\details{
This function launches the Sashimi visualization
R-shiny app.

The key input data should be supplied through argument
\code{envir} which should be prepared using
\code{sashimiDataConstants()}. It will prepare all the
necessary data to use.
\subsection{Custom Parameters}{

The environment \code{envir} may have additional visual
customizations, which recognize the following names:
\itemize{
\item 'panel_height', default \code{200} pixels
\item 'gene_panel_height', default \code{400} pixels
\item 'layout_ncol', default 1
\item 'min_junction_reads', default \code{100}
\item 'share_y_axis', default \code{FALSE}
\item 'use_exon_range', default NULL
\item 'gene_coords_default', default NULL
\item 'label_junctions', default \code{TRUE}
\item 'show_gene_model', default \code{TRUE}
\item 'show_tx_model', default \code{TRUE}
\item 'show_detected_tx', default \code{TRUE}
\item 'font_sizing', default "Default", note it must match available
choices in the UI: "-4 smaller" to "+4 larger".
\item 'exon_font_sizing', default "Default", see comment for
'font-sizing', the same applies.
\item 'junction_alpha', default \code{0.7} for 70\% opacity
\item 'junction_arc_factor', default "Default", note it must match
available choices in the UI: "-2 flat" to "+3 higher".
\item 'junction_arc_minimum', default \code{500}
\item 'aboutExtra', default is html tag data suitable to render
in an HTML section of the R-shiny web page. This entry
is useful to describe the data or experiment underlying
the Splicejam visualizations shown.
}

The environment is prepared by calling \code{sashimiAppConstants()}
which also defines \code{aboutExtra} and other UI elements.
It calls \code{sashimiDataConstants()} which will verify the
environment is usable.
}

\subsection{Running within shiny-server}{

Create a file \code{'app.R'} and include lines which create
the environment, for example:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Prepare data
sjenv <- sashimiDataConstants(gtf=gtf, filesDF=filesDF)

# Add any customizations
sjenv$default_gene <- "Gria1";
sjenv$panel_height <- 100;
sjenv$gene_panel_height <- 200;

# Launch the app
launchSashimiApp()
}\if{html}{\out{</div>}}

This file should be suitable to run the shiny server.

Note that the default will load Farris et al data into the
environment, controlled by \code{empty_uses_farrisdata=TRUE}.
}

\subsection{Other data notes}{

The \code{filesDF} object should be a \code{data.frame} with
at least three colnames:
\itemize{
\item \code{"sample_id"}
\item \code{"type"} (with values either \code{"bw"} or \code{"junction"})
\item \code{"url"} (a URL or file path to each file.)
}

In addition, it recognizes \code{"scale_factor"} with \code{numeric}
values used to scale (by multiplication) the corresponding
scores.

If coverage or junctions are available in multiple files for
the same \code{'sample_id'}, for example when using
replicates, use the same \code{sample_id} for each file, and
the coverage and junctions will be combined by taking the sum.
The optional \code{"scale_factor"} is applied to each file first,
which permits adjusting the files upfront.

The reason for using the sum is that the Splicejam visualization
is intended to show the evidence supporting particular exons
and junctions, and this evidence is built up across replicates.

For more direct control over the data preparation, including
\code{tx2geneDF}, \code{detectedTx}, \code{exonsByGene}, and \code{flatExonsByGene},
see \code{sashimiAppConstants()} which calls \code{sashimiDataConstants()},
both of these functions return an environment that contains the
required data.

When the R-shiny app is created, the \code{ui} and \code{server} components
have their environments set to \code{envir} - so their context will
include the variables defined in that environment.
}

\subsection{Troubleshooting}{
\enumerate{
\item Error \code{"SSL peer certificate or SSH remote key was not OK"} or
\code{"SSL certificate problem: unable to get local issuer certificate"}
seen in the console output of the R shiny app.
\itemize{
\item This error occurs when \code{launchSashimiApp()} references
coverage data on an external web server, for example when
using \code{farrisdata} example data. The remote web server
certificates used with \code{https} web address are using
a certificate whose issuer is not recognized.
\item It means the authority that signed the certificate
is not recognized as an approved authority.
\item There may be two workarounds to this issue:
}
\enumerate{
\item If the certificate authority should be approved, sometimes
on linux hosts it is enough to add certificate extensions,
for example on Ubuntu, or Ubuntu Docker images, one may run

\if{html}{\out{<div class="sourceCode">}}\preformatted{apt-get ca-certificates
# or
sudo apt-get ca-certificates
}\if{html}{\out{</div>}}

Other linux systems may require a different installation.
\item The certificate verification can be skipped temporarily.
Define this option within the same R session:

\if{html}{\out{<div class="sourceCode">}}\preformatted{httr::set_config(httr::config(ssl_verifypeer=FALSE))
}\if{html}{\out{</div>}}

Of course, this option should only be used in trusted
circumstances, for example when accessing a local and trusted
host.
}
\item Error \code{"covNames must be in colnames(GenomicRanges::values(gr))"},
or other error related to \code{names(covNames)} or \code{names(covNamesL)}.
\itemize{
\item These errors ultimately mean there was no available coverage data,
which is usually caused by inability to access the coverage
file itself. The web server may be offline, or may be denying
connection. The path to the file may be incorrect.
\item Typically the coverage data is obtained from memoise cached data,
and only when the cache is not available, or somehow incorrect,
the data is retrieved from the file or remote server.
}
}
}
}
\examples{
# Note: disabled for web page examples
# launchSashimiApp();

}
\seealso{
Other Splicejam core functions:
\code{\link[=sashimiDataConstants]{sashimiDataConstants()}},
\code{\link[=splicejamDataFromTxDb]{splicejamDataFromTxDb()}},
\code{\link[=splicejamFigure]{splicejamFigure()}}
}
\concept{Splicejam core functions}
