% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{sashimiPrep}
\alias{sashimiPrep}
\title{Prepare Sashimi plot}
\usage{
sashimiPrep(inputGenes, exonsGR, expandUpDown = c(0, 0),
  WHICHEXONS = NULL, YMAX = NULL, compressGaps = TRUE,
  GAPWIDTH = 100, readCountCex = 1.5, readCountFont = 1,
  readCountSrt = 0, maxLwd = 5, readColor = "#BB000099",
  readCountPadding = "         ", bigWigFiles = NULL, covL = NULL,
  spliceCountsDF = NULL, colorFunc2 = function(n, ...) {    
  head(rainbowJam(max(c(n, 5))), n) }, cex.main = 1.7, font.main = 2,
  alternateSpliceUpDown = FALSE, srtSplice = 0, arcScale = 20,
  arcScaleFactor = 100, useAspect = 2, transformY = c("none", "sqrt",
  "log2"), unionCoverages = FALSE, unionLabel = "unionCoverage",
  ylimFactor = 1.5, spliceGRs = NULL, spliceBuffer = 3,
  exactJunctionsOnly = TRUE, absValues = TRUE,
  countAsPercent = FALSE, geneSymbolColname = c("geneSymbol",
  "geneSymbolChrom", "gene_id", "gene_name"), flipStrand = TRUE,
  ignoreSpliceStrand = TRUE, spliceOverlapType = c("within", "any",
  "start", "end"), reverseStrandCoords = TRUE, showCoords = TRUE,
  labelExons = FALSE, labelExonName = "exonLabel", doPar = TRUE,
  parLend = "round", doTitle = TRUE, doMargins = TRUE,
  doPlot = TRUE, minReads = 0, VERBOSE = TRUE,
  exonLabelColname = "gene_nameExon", verbose = VERBOSE, ...)
}
\description{
Prepare Sashimi plot
}
