% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{shrinkMatrix}
\alias{shrinkMatrix}
\title{Shrink numeric matrix by groups of rows}
\usage{
shrinkMatrix(
  x,
  groupBy,
  shrinkFunc = function(x) {
     .Internal(mean(x))
 },
  returnClass = c("data.frame", "matrix"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric matrix}

\item{groupBy}{vector of group labels, whose length equals \code{nrow(x)}.
These values will become rownames in the output data.}

\item{shrinkFunc}{function that takes vector input and returns
vector output. The vector class can be checked, in order to
call a function on numeric or character data separately, as
needed.}

\item{returnClass}{character string indicating the return data type,
\code{"data.frame"} returns a \code{data.frame} whose first column contains
entries from \code{groups}; \code{"matrix"} returns a numeric matrix whose
rownames are entries from \code{groups}.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\description{
Shrink numeric matrix by groups of rows
}
\details{
This function is mainly a wrapper around the very efficient
functions in the \code{data.table} package, with the ability to
provide a custom function to shrink row values.
}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
