% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{simplifyXY}
\alias{simplifyXY}
\title{Simplify XY coordinates to minimal line segments}
\usage{
simplifyXY(xy, minN = 3, restrictDegrees = NULL, ...)
}
\arguments{
\item{xy}{numeric matrix of two columns x,y}

\item{minN}{integer minimum number of consecutive points to
cause compression to occur. It requires at least 3 points
inherent to the algorithm.}

\item{restrictDegrees}{numeric vector of allowed angles in degrees
(range 0 to 360) of angles allowed to be compressed, when supplied
all other angles will not be compressed.}

\item{...}{additional arguments are ignored.}
}
\description{
Simplify XY coordinates to minimal line segments
}
\details{
This function takes a numeric matrix of x,y coordinates
and returns the minimal matrix of x,y coordinates that represents
the same line segments. It is intended in cases where there
is a long repeated line segment that could be represented by
far fewer points.
}
\examples{
xy <- cbind(
   x=c(1,1:15,1),
   y=c(0, 0,0,0,0,1,2,3,4,5,5,5,5,6,0,0, 0));
par("mfrow"=c(1,1));
plot(xy, cex=2);
points(simplifyXY(xy), pch=20, col="orange", cex=3);
points(simplifyXY(xy, restrictDegrees=c(0,180)), pch=17, col="purple");
legend("topleft", pch=c(1, 20, 17),# box.col="black", border="black",
   pt.cex=c(2, 3, 1),
   col=c("black", "orange", "purple"),
   legend=c("all points", "simplified points", "only simplified horizontal"));

}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=stackJunctions]{stackJunctions()}},
\code{\link[=strsplitOrdered]{strsplitOrdered()}}
}
\concept{Internal utility functions}
