% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{sortGRL}
\alias{sortGRL}
\title{Sort GRangesList elements by chromosome and position}
\usage{
sortGRL(
  GRL,
  splitColname = "splitColname",
  keep_order = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{GRL}{GRangesList to be sorted.}

\item{splitColname}{intermediate colname used to split values
from GRanges to GRangesList. It only needs to be defined
if for some reason the default "splitColname" is already
in \code{colnames(GenomicRanges::values(GRL))}.}

\item{keep_order}{logical indicating whether to maintain the
input order of GRanges; \code{FALSE} will return GRangesList
ordered by the first sorted GRanges occurrence.}

\item{...}{additional arguments are ignored.}
}
\value{
GRangesList sorted by chromosome and position.
}
\description{
Sort GRangesList elements by chromosome and position
}
\details{
This function sorts GRangesList elements by chromosome and
position, using the default sort routine for GRanges objects.
It accomplishes the task by sorting the GRanges elements, then
splitting that GRanges into a GRangesList based upon the
original \code{names(GrangesList)}.
}
\examples{
gr12 <- GenomicRanges::GRanges(
   seqnames=rep(c("chr1", "chr2", "chr1"), c(3,3,3)),
   ranges=IRanges::IRanges(
      start=c(100, 200, 400, 500, 300, 100, 200, 400, 600),
      width=c(50,50,50, 50,50,50, 50,50,50)
   ),
   strand=rep(c("+", "-", "+"), c(3,3,3)),
   gene_name=rep(c("GeneA", "GeneB", "GeneC"), each=3)
)
grl <- GenomicRanges::split(gr12, GenomicRanges::values(gr12)$gene_name);
grl;
GenomicRanges::sort(grl);
sortGRL(grl);
sortGRL(grl, keep_order=FALSE);

}
\seealso{
Other GenomicRanges functions:
\code{\link[=addGRLgaps]{addGRLgaps()}},
\code{\link[=addGRgaps]{addGRgaps()}},
\code{\link[=annotateGRLfromGRL]{annotateGRLfromGRL()}},
\code{\link[=annotateGRfromGR]{annotateGRfromGR()}},
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=closestExonToJunctions]{closestExonToJunctions()}},
\code{\link[=findOverlapsGRL]{findOverlapsGRL()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getGRLgaps]{getGRLgaps()}},
\code{\link[=getGRgaps]{getGRgaps()}},
\code{\link[=jam_isDisjoint]{jam_isDisjoint()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}}
}
\concept{GenomicRanges functions}
