% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-design.R
\name{strsplitOrdered}
\alias{strsplitOrdered}
\title{Split the elements of an ordered factor vector}
\usage{
strsplitOrdered(
  x,
  split = "_",
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  sortFunc = jamba::mixedSort,
  keepOrder = TRUE,
  ...
)
}
\arguments{
\item{x}{character or factor vector.}

\item{split}{character split value sent to \code{base::strsplit()}.}

\item{fixed, perl, useBytes}{additional arguments sent to \code{base::split()}.}

\item{sortFunc}{function used to sort character values when the input
\code{x} is a character vector. The default \code{jamba::mixedSort()} applies
alphanumeric sort.}

\item{keepOrder}{logical indicating whether to keep the order of values
in the input data, for example with character input the values will
be ordered by the first appearance of each term.}

\item{...}{additional arguments are ignored.}
}
\value{
list of factor vectors, where each factor shares the same
global factor levels based upon the input data.
}
\description{
Split the elements of an ordered factor vector
}
\details{
This function performs \code{base::strsplit()} while trying to maintain
the order of factor levels in the output, based upon the order of
factor levels in the input data.
}
\examples{
# first define a vector of sample groups
iGroups <- jamba::nameVector(paste(rep(c("WT", "KO"), each=6),
   rep(c("Control", "Treated"), each=3),
   sep="_"));
iGroups <- factor(iGroups, levels=unique(iGroups));
iGroups;
strsplitOrdered(iGroups, "_");

}
\seealso{
Other Internal utility functions:
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=compressPolygonM]{compressPolygonM()}},
\code{\link[=df2colorSub]{df2colorSub()}},
\code{\link[=dfWide2segments]{dfWide2segments()}},
\code{\link[=escapeWhitespaceRegexp]{escapeWhitespaceRegexp()}},
\code{\link[=factor2label]{factor2label()}},
\code{\link[=geomean]{geomean()}},
\code{\link[=intercalate]{intercalate()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=jamGeomean]{jamGeomean()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=shrinkMatrix]{shrinkMatrix()}},
\code{\link[=simplifyXY]{simplifyXY()}},
\code{\link[=stackJunctions]{stackJunctions()}}
}
\concept{Internal utility functions}
