% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gg.R
\name{to_basic.GeomShape}
\alias{to_basic.GeomShape}
\title{Support plotly for GeomShape}
\usage{
\method{to_basic}{GeomShape}(data, prestats_data, layout, params, p, ...)
}
\arguments{
\item{data}{the data returned by \code{ggplot2::ggplot_build()}.}

\item{prestats_data}{the data before statistics are computed.}

\item{layout}{the panel layout.}

\item{params}{parameters for the geom, statistic, and 'constant' aesthetics}

\item{p}{a ggplot2 object (the conversion may depend on scales, for instance).}

\item{...}{currently ignored}
}
\description{
Support plotly for GeomShape
}
\details{
This function is a helper function intended to provide a simple
but not fully-equivalent link between \code{ggforce::geom_shape()} and
\code{plotly::ggplotly()}. This function converts data to
\code{ggplot2::geom_polygon} which is sufficient for the splicejam
package, but which does not provide the extra effects from
\code{ggforce::geom_shape()} such as rounded corners or resizing.
This function may be extended then submitted to
the \code{ggforce} package to assist plotly support.

This function is exported, since otherwise it caused problems
when invoking \code{plotly::ggplotly()} and this function was not
consistently used in that process for some reason.

Currently only \code{data} is passed to \code{ggplot2::GeomPolygon}.
}
\seealso{
Other ggplot2 customizations:
\code{\link[=geom_diagonal_wide_arc]{geom_diagonal_wide_arc()}},
\code{\link{splicejam-extensions}},
\code{\link[=stat_unpack_polygon]{stat_unpack_polygon()}}
}
\concept{ggplot2 customizations}
