% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{addGRLgaps}
\alias{addGRLgaps}
\title{Add gaps between GRangesList regions}
\usage{
addGRLgaps(grl, strandSpecific = TRUE, gapname = "gap",
  suffix = "_v", newValues = list(feature_type = "gap"),
  doSort = TRUE, ...)
}
\arguments{
\item{grl}{GRangesList object}

\item{strandSpecific}{logical indicating whether the gaps are calculated
per strand, see \code{getGRLgaps()}.}

\item{gapname, suffix}{character vector supplying the name to assign to new
gap GRanges elements, using \code{jamba::makeNames()} with \code{suffix} as
described to define non-duplicated names. If \code{gapname is NULL} then
no names are assigned to new gap GRanges entries, however when the
input \code{gr} GRanges object has names, the concatenation of gaps
causes names \code{""} to be assigned to all gap GRanges elements, which
are duplicated for multiple gaps.}

\item{newValues}{list of values to add to the resulting gap GRanges,
whose names become \code{colnames(grl@unlistData)}, and whose values are used
to populate each column. By default a colname \code{"feature_type"} is
added, with value \code{"gap"} added to each row. When \code{newValues is NULL}
then no values are added to the gaps GRanges.}

\item{doSort}{logical indicating whether to sort the resulting
GRanges objects. When \code{doSort=FALSE} the gaps are added to the end
of each input \code{grl} GRanges object. Note that the GrangesList object
is not sorted, only the GRanges objects within the GRangesList
are sorted.}

\item{...}{additional arguments are passed to \code{getGRLgaps()}.}
}
\value{
GRangesList object, sorted per GRangesList element
when \code{doSort=TRUE}. When \code{newValues}
is supplied, the values for gaps GRanges elements will be assigned,
otherwise any column values present in \code{gr} will be \code{NA} for
gaps elements. The names of gaps elements are assigned using
\code{gapname} then are made unique using \code{jamba::makeNames()},
unless \code{gapname is NULL}.
}
\description{
Add gaps between GRangesList regions
}
\details{
This function adds gaps between each GRanges region separately
for each GRangesList element, where
there is a gap between two GRanges for
the same seqnames. When \code{strandSpecific=TRUE} the gaps are
determined per strand.

This function is a wrapper around \code{getGRLgaps()}, which is then
concatenated to the input \code{gr} GRanges object using \code{S4Vectors::pc()}.
When the input \code{grl} GRanges has column \code{S4Vectors::values()} then the
gaps GRanges object will have \code{NA} values used by default. To supply
values, use the \code{newValues} argument, which assigns name-value pairs.
}
\examples{
gr <- GRanges(seqnames=rep(c("chr1","chr2"), c(3,2)),
   ranges=IRanges(start=c(100, 300, 400, 300, 700),
      end=c(199, 450, 500, 600, 800)),
   strand=rep(c("+","-"), c(3,2)),
   feature_type="exon");
names(gr) <- makeNames(rep("exon", length(gr)));
gr;
addGRgaps(gr);

grl <- GenomicRanges::split(gr, seqnames(gr));
grl;
addGRLgaps(grl);
addGRLgaps(grl, strandSpecific=FALSE);

}
\seealso{
Other jam GRanges functions: \code{\link{addGRgaps}},
  \code{\link{annotateGRLfromGRL}},
  \code{\link{annotateGRfromGR}},
  \code{\link{assignGRLexonNames}},
  \code{\link{findOverlapsGRL}},
  \code{\link{getFirstStrandedFromGRL}},
  \code{\link{sortGRL}}
}
\concept{jam GRanges functions}
