% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{bgaPlotly3d}
\alias{bgaPlotly3d}
\title{Create an interactive 3-D BGA plotly visualization}
\usage{
bgaPlotly3d(
  bgaInfo,
  axes = c(1, 2, 3),
  superGroups = NULL,
  superGroupAlpha = 0.3,
  superGroupLwd = 25,
  arrowSmoothFactor = 8,
  colorSub = NULL,
  drawVectors = c("none", "centroids", "genes"),
  highlightGenes = NULL,
  drawSampleLabels = TRUE,
  ellipseType = c("ellipsoid", "alphahull", "none"),
  ellipseOpacity = 0.2,
  useScaledCoords = FALSE,
  geneColor = "#555555",
  geneScaleFactor = 1,
  geneLwd = 2,
  geneAlpha = 0.7,
  geneLabels = NULL,
  maxGenes = 50,
  centroidLwd = 15,
  centroidAlpha = 0.3,
  textposition_centroid = "top center",
  textposition_sample = "middle right",
  sceneX = -1.25,
  sceneY = 1.25,
  sceneZ = 1.25,
  main = "",
  plot_bgcolor = "#eeeeee",
  paper_bgcolor = "#dddddd",
  sampleGroups = NULL,
  debug = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bgaInfo}{object of class \code{c("coa", "bga")} created by \code{made4::bga()}.}

\item{axes}{\code{integer} vector length 3 indicating the BGA axes to use
for the 3-D visualization. It is sometimes helpful to use
\code{axes=c(2,3,4)} especially when the first component suggests
a large experimental batch effect, for example when
combining data from multiple GEO studies, often the first
BGA component is indicative of the different experimental
platforms.}

\item{superGroups}{\code{character} or factor vector with length
\code{length(bgaInfo$fac)}, in the same order. A super group consists
of multiple sample groups, as defined by values in \code{bgaInfo$fac}.}

\item{superGroupAlpha}{\code{numeric} value between 0 and 1 indicating the
alpha transparency to use for the vector that connects groups
within each super group.}

\item{superGroupLwd}{\code{numeric} line width when connecting sample group
centroids by \code{superGroups}.}

\item{arrowSmoothFactor}{\code{numeric} value typically ranging from 1 to 5,
affecting the amount of curve smoothing when generating a
3-D spline curve to connect more than two super groups.
When there are only two group members in one super group,
the two groups will always use a straight line. For three or
more groups, a 3-D spline is used, in order to help visualize
trends by means of a curve.}

\item{colorSub}{\code{character} vector of R colors, whose names contain values in
\code{bgaInfo$fac}. These colors are used for the samples, sample
centroids, and are shaded via a gradient when connecting multiple
centroids using \code{superGroups}.}

\item{drawVectors}{\code{character} vector of one or more types of vectors:
\itemize{
\item \code{"centroids"} draws a vector from the origin to each group centroid.
\item \code{"genes"} draws a vector from the origin to each gene.
}}

\item{drawSampleLabels}{\code{logical} indicating whether to label each
individual sample point, as opposed to labeling only the
sample group centroid.}

\item{ellipseType}{\code{character} value indicating how to draw sample
centroid ellipses, where \code{"none"} draws no ellipse; \code{"alphahull"}
draws a wireframe; and \code{"ellipsoid"} draws
a partially transparent shell. Note that plotly does not handle
transparency well in 3-D space, sometimes fully obscuring
background components.}

\item{ellipseOpacity}{\code{numeric} value between 0 and 1 indicating the
transparency of the ellipse, when \code{ellipseType} is not \code{"none"};
0 is fully transparent, and 1 is non-transparent.}

\item{useScaledCoords}{\code{logical} indicating whether to use the BGA
scaled coordinates, or the raw coordinates. By default the raw
coordinates are used in order to preserve the relative contribution
of each BGA component to the overall variability (or inertial
moment.)}

\item{geneColor}{\code{character} color used when displaying gene points.}

\item{geneScaleFactor}{\code{numeric} value used to expand the position of
genes relative to the origin, intended to help visualize the vector
position of genes relative to the origin.}

\item{geneLwd, geneAlpha}{\code{numeric} used to customize the line width
and alpha transparency of gene points, respectively.}

\item{geneLabels}{\code{character} vector of gene labels to use instead
of rownames. The \code{names(geneLabels)} are expected to match
\code{rownames(bgaInfo$bet$co)}. This option is intended to allow
display of gene symbols, instead of an assay identifier like
probe set name.}

\item{maxGenes}{\code{integer} maximum number of genes to display.}

\item{centroidLwd, centroidAlpha}{\code{numeric} used to customize the
line width and alpha transparency of vectors drawn to sample
centroids, respectively.}

\item{textposition_centroid, textposition_sample}{\code{character} vector
containing values used as the \code{plotly} argument \code{"textposition"}
to position text labels for centroids, and samples. Each
vector is extended to the number of labels, which allows
each label to be individually positioned.}

\item{sceneX, sceneY, sceneZ}{\code{numeric} used to define the camera
position in 3-D space as defined by plotly. Frankly, the ability
to customize the starting position is fairly confusing, but these
values are faithfully passed along to the corresponding plotly
call.}

\item{main}{\code{character} string used as a title on the plotly
visualization.}

\item{sampleGroups}{\code{character} or \code{factor} vector, whose
names are expected to contain values in
\code{rownames(bgaInfo$ord$ord$tab)}, which should also be \code{colnames(x)}
of the input data to \code{made4::bga(x, ...)}. The values in
\code{sampleGroups} are used to re-group samples, which effectively
calculates new sample group centroids used in this
visualization. It is intended to be used when the \code{made4::bga()}
is used without grouping, for example when each sample replicate
is defined as its own group. The end result is that centroids
and ellipsoids are calculated for each group during visualization,
to help organize the points displayed.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional parameters are ignored}
}
\value{
plotly object sufficient to render an HTML page containing
a 3-D BGA plot.
}
\description{
Create an interactive 3-D BGA plotly visualization
}
\details{
This function takes the output from \code{made4::bga()} (class \code{"bga"})
and creates a 3D plotly visualization which shows the samples,
sample centroids, and optionally a subset of genes in the same
BGA component space.

For general guidance, when the data contains four or more sample
groups, this function can be called to produce a 3-D plot.
BGA itself produces \code{n-1} dimensions, therefore when there are
fewer than four groups, a 3-D plot cannot be produced.

When analyzing data with fewer than four groups, it is sometimes
helpful to cluster samples individually. This approach is roughly
equivalent to performing PCA or COA directly, but is sometimes
convenient in being consistent with calling \code{made4::bga()}, as
it also applies the same scaling method used in \code{made4:bga()}
prior to calling PCA or COA. In that case, the groups can still
be displayed by \code{bgaPlotly3d()} by defining the \code{sampleGroups}
argument. The result will display typical PCA or COA coordinates,
but visually displays the centroid of each sample group.

This function is especially useful when there are multiple
sample groups that belong to a "supergroup" -- a higher level
of grouping. Consider an experiment with two cell lines treated
over multiple time points. The samples would be assigned to
a group based upon cell_line and time, then each group would
be assigned to a supergroup based upon cell_line. The output
plot will display each sample group, group centroid, then
connect group centroids in order by the supergroup.
The order of groups within each supergroup is determined
either by factor levels, or by the order returned by
\code{jamba::mixedSort()} which performs a proper alphanumeric sort.
\subsection{Background on \code{bga} object content}{

Output from \code{made4::bga()} is an object with \code{class} containing \code{"bga"}.
It is a list with three elements:
\enumerate{
\item \code{"ord"}: ordination, the initial clustering results from either \code{"coa"}
(default) or \code{"pca"} clustering across all individual samples. This data
is fed into the second phase, group-based clustering.
\item \code{"bet"}: the between group analysis result, derived from \code{"ord"}.
\item \code{"fac"}: \code{factor} vector of groups across the \code{colnames(x)} of
input data.
}

The \code{"bet"} contents are described below, assuming the results of
\code{made::bga()} are stored as \code{bgaInfo}:
\itemize{
\item gene (row) coordinates are stored:
\itemize{
\item \code{bgaInfo$bet$co}: unscaled coordinates for each component axis
\item \code{bgaInfo$bet$c1}: scaled coordinates for each component axis
}
\item group centroid coordinates are stored:
\itemize{
\item \code{bgaInfo$bet$li}: unscaled group centroid coordinates
\item \code{bgaInfo$bet$ls}: scaled group centroid coordinates
}
\item sample centroid coordinates are stored:
\itemize{
\item \code{bgaInfo$bet$ls}: scaled coordinates for each component axis
}
}
}
}
\seealso{
Other jam plot functions: 
\code{\link{factor2label}()},
\code{\link{gene2gg}()},
\code{\link{grl2df}()},
\code{\link{jitter_norm}()},
\code{\link{plotSashimi}()},
\code{\link{prepareSashimi}()},
\code{\link{stackJunctions}()}

Other jam spatial functions: 
\code{\link{compressPolygonM}()},
\code{\link{dfWide2segments}()},
\code{\link{simplifyXY}()},
\code{\link{spline3d}()}
}
\concept{jam plot functions}
\concept{jam spatial functions}
