% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{bgaPlotly3d}
\alias{bgaPlotly3d}
\title{Create an interactive 3-D BGA plotly visualization}
\usage{
bgaPlotly3d(bgaInfo, axes = c(1, 2, 3), superGroups = NULL,
  superGroupAlpha = 0.3, superGroupLwd = 25, arrowSmoothFactor = 8,
  colorSub = NULL, drawVectors = c("none", "centroids", "genes"),
  highlightGenes = NULL, drawSampleLabels = TRUE,
  ellipseType = c("ellipsoid", "alphahull", "none"),
  useScaledCoords = FALSE, geneColor = "#555555",
  geneScaleFactor = 1, geneLwd = 2, geneAlpha = 0.7, maxGenes = 50,
  centroidLwd = 15, centroidAlpha = 0.3, sceneX = -1.25,
  sceneY = 1.25, sceneZ = 1.25, main = "",
  plot_bgcolor = "#eeeeee", paper_bgcolor = "#dddddd", debug = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{bgaInfo}{object of class \code{"bga"} as created by \code{made4::bga()}.}

\item{axes}{integer vector length 3 indicating the BGA axes to use
for the 3-D visualization. It is sometimes helpful to use
\code{axes=c(2,3,4)} especially when the first component suggests
a large experimental batch effect, for example when
combining data from multiple GEO studies, often the first
BGA component is indicative of the different experimental
platforms.}

\item{superGroups}{character or factor vector with length
\code{length(bgaInfo$fac)}, in the same order. A super group consists
of multiple sample groups, as defined by values in \code{bgaInfo$fac}.}

\item{superGroupAlpha}{numeric value between 0 and 1 indicating the
alpha transparency to use for the vector that connects groups
within each super group.}

\item{superGroupLwd}{the line width when connecting sample group
centroids by \code{superGroups}.}

\item{arrowSmoothFactor}{numeric value typically ranging from 1 to 5,
affecting the amount of curve smoothing when generating a
3-D spline curve to connect more than two super groups.
When there are only two group members in one super group,
the two groups will always use a straight line. For three or
more groups, a 3-D spline is used, in order to help visualize
trends by means of a curve.}

\item{colorSub}{named color vector, whose names contain values in
\code{bgaInfo$fac}. These colors are used for the samples, sample
centroids, and are shaded via a gradient when connecting multiple
centroids using \code{superGroups}.}

\item{drawVectors}{character vector of one or more types of vectors,
where \code{"centroids"} draws a vector from the origin to each group
centroid; and \code{"genes"} draws a vector from the origin to each
gene.}

\item{drawSampleLabels}{logical indicating whether to label each
individual sample point, as opposed to labeling only the
sample group centroid.}

\item{useScaledCoords}{logical indicating whether to use the BGA
scaled coordinates, or the raw coordinates. By default the raw
coordinates are used in order to preserve the relative contribution
of each BGA component to the overall variability (or inertial
moment.)}

\item{geneColor}{character color used when displaying gene points.}

\item{geneScaleFactor}{numeric value used to expand the position of
genes relative to the origin, intended to help visualize the vector
position of genes relative to the origin.}

\item{geneLwd, geneAlpha}{parameters used to customize the line width
and alpha transparency of gene points, respectively.}

\item{maxGenes}{integer maximum number of genes to display.}

\item{centroidLwd, centroidAlpha}{parameters used to customize the
line width and alpha transparency of vectors drawn to sample
centroids, respectively.}

\item{sceneX, sceneY, sceneZ}{parameters used to define the camera
position in 3-D space as defined by plotly. Frankly, the ability
to customize the starting position is fairly confusing, but these
values are faithfully passed along to the corresponding plotly
call.}

\item{main}{character string used as a title on the plotly
visualization.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters are ignored}

\item{elipseType}{character value indicating how to draw sample
centroid ellipses, where \code{"none"} draws no ellipse; \code{"alphahull"}
draws a wireframe; and \code{"ellipsoid"} draws
a partially transparent shell. Note that plotly does not handle
transparency well in 3-D space, sometimes fully obscuring
background components.}
}
\description{
Create an interactive 3-D BGA plotly visualization
}
\details{
This function takes the output from \code{made4::bga()} (class \code{"bga"})
and creates a 3D plotly visualization which shows the samples,
sample centroids, and optionally a subset of genes in the same
BGA component space.
}
\concept{jam plot functions}
