% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{closestExonToJunctions}
\alias{closestExonToJunctions}
\title{Find closest exon to splice junction ends}
\usage{
closestExonToJunctions(
  spliceGRgene,
  exonsGR,
  flipNegativeStrand = TRUE,
  sampleColname = "sample_id",
  reportActualCoords = FALSE,
  spliceBuffer = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{spliceGRgene}{GRanges representing splice junctions}

\item{exonsGR}{GRanges representing flattened exons per gene, as
is produced by \code{flattenExonsBy()}.}

\item{flipNegativeStrand}{logical indicating whether to flip
the orientation of features on the negative strand.}

\item{sampleColname}{character value matching the colname that
defines distinct sample identifier, for which junctions will
be kept separate.}

\item{reportActualCoords}{logical indicating whether to report
genomic coordinates or transcriptome coordinates. (Work in
progress.)}

\item{spliceBuffer}{optional \code{integer} indicating the maximum distance
from an exon in order for an exon to be assigned to the exon.
When \code{spliceBuffer=NULL} the distance is ignored in value
columns \code{c("nameFrom","nameTo")}. When \code{spliceBuffer} is supplied,
and junction is greater than this distance, the value columns
\code{c("nameFrom","nameTo")} will indicate the exon name appended
to the distance.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Find closest exon to splice junction ends
}
\details{
This function is used to annotate splice junction GRanges entries
based upon the closest compatible stranded exon boundary.

This function should usually be called by \code{spliceGR2junctionDF()}
and not called directly.
}
\seealso{
Other jam RNA-seq functions:
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=defineDetectedTx]{defineDetectedTx()}},
\code{\link[=detectedTxInfo]{detectedTxInfo()}},
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getGRcoverageFromBw]{getGRcoverageFromBw()}},
\code{\link[=groups2contrasts]{groups2contrasts()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=makeTx2geneFromGtf]{makeTx2geneFromGtf()}},
\code{\link[=make_ref2compressed]{make_ref2compressed()}},
\code{\link[=prepareSashimi]{prepareSashimi()}},
\code{\link[=runDiffSplice]{runDiffSplice()}},
\code{\link[=sortSamples]{sortSamples()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}}

Other jam GRanges functions:
\code{\link[=addGRLgaps]{addGRLgaps()}},
\code{\link[=addGRgaps]{addGRgaps()}},
\code{\link[=annotateGRLfromGRL]{annotateGRLfromGRL()}},
\code{\link[=annotateGRfromGR]{annotateGRfromGR()}},
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=findOverlapsGRL]{findOverlapsGRL()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getFirstStrandedFromGRL]{getFirstStrandedFromGRL()}},
\code{\link[=getGRLgaps]{getGRLgaps()}},
\code{\link[=getGRcoverageFromBw]{getGRcoverageFromBw()}},
\code{\link[=getGRgaps]{getGRgaps()}},
\code{\link[=grl2df]{grl2df()}},
\code{\link[=jam_isDisjoint]{jam_isDisjoint()}},
\code{\link[=make_ref2compressed]{make_ref2compressed()}},
\code{\link[=sortGRL]{sortGRL()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}},
\code{\link[=stackJunctions]{stackJunctions()}}
}
\concept{jam GRanges functions}
\concept{jam RNA-seq functions}
