% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{closestExonToJunctions}
\alias{closestExonToJunctions}
\title{Find closest exon to splice junction ends}
\usage{
closestExonToJunctions(
  spliceGRgene,
  exonsGR,
  flipNegativeStrand = TRUE,
  sampleColname = "sample_id",
  reportActualCoords = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{spliceGRgene}{GRanges representing splice junctions}

\item{exonsGR}{GRanges representing flattened exons per gene, as
is produced by \code{flattenExonsBy()}.}

\item{flipNegativeStrand}{logical indicating whether to flip
the orientation of features on the negative strand.}

\item{sampleColname}{character value matching the colname that
defines distinct sample identifier, for which junctions will
be kept separate.}

\item{reportActualCoords}{logical indicating whether to report
genomic coordinates or transcriptome coordinates. (Work in
progress.)}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Find closest exon to splice junction ends
}
\details{
This function is used to annotate splice junction GRanges entries
based upon the closest compatible stranded exon boundary.

This function should usually be called by \code{spliceGR2junctionDF()}
and not called directly.
}
\seealso{
Other jam RNA-seq functions: 
\code{\link{assignGRLexonNames}()},
\code{\link{combineGRcoverage}()},
\code{\link{defineDetectedTx}()},
\code{\link{detectedTxInfo}()},
\code{\link{exoncov2polygon}()},
\code{\link{flattenExonsBy}()},
\code{\link{groups2contrasts}()},
\code{\link{makeTx2geneFromGtf}()},
\code{\link{make_ref2compressed}()},
\code{\link{runDiffSplice}()},
\code{\link{sortSamples}()},
\code{\link{spliceGR2junctionDF}()}

Other jam GRanges functions: 
\code{\link{addGRLgaps}()},
\code{\link{addGRgaps}()},
\code{\link{annotateGRLfromGRL}()},
\code{\link{annotateGRfromGR}()},
\code{\link{assignGRLexonNames}()},
\code{\link{combineGRcoverage}()},
\code{\link{exoncov2polygon}()},
\code{\link{findOverlapsGRL}()},
\code{\link{getFirstStrandedFromGRL}()},
\code{\link{getGRLgaps}()},
\code{\link{getGRcoverageFromBw}()},
\code{\link{getGRgaps}()},
\code{\link{grl2df}()},
\code{\link{make_ref2compressed}()},
\code{\link{sortGRL}()},
\code{\link{spliceGR2junctionDF}()},
\code{\link{stackJunctions}()}
}
\concept{jam GRanges functions}
\concept{jam RNA-seq functions}
