% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{closestExonToJunctions}
\alias{closestExonToJunctions}
\title{Find closest exon to splice junction ends}
\usage{
closestExonToJunctions(spliceGRgene, exonsGR, flipNegativeStrand = TRUE,
  sampleColname = sampleColname, reportActualCoords = FALSE,
  verbose = TRUE, ...)
}
\description{
Find closest exon to splice junction ends
}
