% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{codonUsage2df}
\alias{codonUsage2df}
\title{Make codon usage data.frame}
\usage{
codonUsage2df(file, ...)
}
\arguments{
\item{file}{path to the codon usage file}
}
\value{
data.frame containing codon usage data, with colnames
\code{codon}, \code{frequency}, and \code{count}.
}
\description{
Make codon usage data.frame
}
\details{
This function imports a text file containing codon usage, and
creates a data.frame with columns containing the \code{codon},
\code{fraction}, and \code{count}. The colname \code{fraction} represents
the fraction observed counts per codon
compared to the codon with the highest counts, for each
amino acid. For example, an amino acid with only one codon
would be \code{c(1.0)}. An amino acid with two codons, with observed
counts \code{c(100, 200)}, would have fraction values \code{c(0.5, 1.0)}.

The format is derived from published files, one example is
included in the package \code{extdata/Mouse_codon_usage.txt},
which can be accessed using the command:

\code{system.file("extdata", "Mouse_codon_usage.txt", package="farrisdata")}
}
\examples{
codonFile <- system.file("extdata", "Mouse_codon_usage.txt", package="farrisdata");
codonDF <- codonUsage2df(codonFile);

}
\seealso{
Other jam RNA-seq functions: \code{\link{defineDetectedTx}},
  \code{\link{detectedTxInfo}}, \code{\link{dna2codon}},
  \code{\link{jamCai}}, \code{\link{makeTx2geneFromGtf}},
  \code{\link{tx2ale}}
}
\concept{jam RNA-seq functions}
