% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{combineGRcoverage}
\alias{combineGRcoverage}
\title{Combine GRanges coverage replicates}
\usage{
combineGRcoverage(gr, covNames = NULL, covName = NULL,
  strands = NULL, scaleFactors = 1, verbose = FALSE, ...)
}
\description{
Combine GRanges coverage replicates
}
\details{
This function takes a GRanges object as output from
\code{getGRcoverageFromBw()} and combines the coverages into
one coverage per strand. The strand is inferred by the
presence of negative values, where any negative value
indicates the column is negative strand.
}
