% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{combineGRcoverage}
\alias{combineGRcoverage}
\title{Combine GRanges coverage replicates}
\usage{
combineGRcoverage(
  gr,
  covNames = NULL,
  covName = NULL,
  strands = NULL,
  scaleFactors = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gr}{GRanges object containing coverage data in columns
containing NumericList class data.}

\item{covNames}{character vector of \code{colnames(values(gr))}
representing columns in \code{gr} that contain coverage data
in NumericList format, for example data prepared
with \code{getGRcoverageFromBw()}.}

\item{covName}{character vector with length equal to
\code{length(covNames)} representing the \code{sample_id} for each
\code{covNames} entry.}

\item{strands}{character vector, or NULL, indicating the strand
for which the coverage data was obtained. When \code{NULL} the strand
is inferred by the presence of any negative values.}

\item{scaleFactors}{numeric vector length equal to \code{length(covNames)}
or expanded to that length. Values are multiplied by each
coverage data result, intended to apply a normalization
to each coverage value. A \code{-1} value can also be used
to flip the score of negative strand data, in the event the
source coverage data is scored only using positive values.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
GRanges object whose colnames contain the
\code{covName} (\code{sample_id}) for each observed strand,
with coverage combined taking the sum of individual
coverages after multiplying each by \code{scaleFactors}.
}
\description{
Combine GRanges coverage replicates
}
\details{
This function takes a GRanges object as output from
\code{getGRcoverageFromBw()} and combines the coverages into
one coverage per strand for each \code{covName} (equivalent
to \code{sample_id}). Each coverage value is multiplied by
its \code{scaleFactors} value, then the sum is returned for
each strand, for each \code{covName} (\code{sample_id}).

The strand is inferred by the
presence of negative values, where any negative value
indicates the column is negative strand.
}
\seealso{
Other jam GRanges functions: 
\code{\link{addGRLgaps}()},
\code{\link{addGRgaps}()},
\code{\link{annotateGRLfromGRL}()},
\code{\link{annotateGRfromGR}()},
\code{\link{assignGRLexonNames}()},
\code{\link{closestExonToJunctions}()},
\code{\link{exoncov2polygon}()},
\code{\link{findOverlapsGRL}()},
\code{\link{getFirstStrandedFromGRL}()},
\code{\link{getGRLgaps}()},
\code{\link{getGRcoverageFromBw}()},
\code{\link{getGRgaps}()},
\code{\link{grl2df}()},
\code{\link{jam_isDisjoint}()},
\code{\link{make_ref2compressed}()},
\code{\link{sortGRL}()},
\code{\link{spliceGR2junctionDF}()},
\code{\link{stackJunctions}()}

Other jam RNA-seq functions: 
\code{\link{assignGRLexonNames}()},
\code{\link{closestExonToJunctions}()},
\code{\link{defineDetectedTx}()},
\code{\link{detectedTxInfo}()},
\code{\link{exoncov2polygon}()},
\code{\link{flattenExonsBy}()},
\code{\link{groups2contrasts}()},
\code{\link{makeTx2geneFromGtf}()},
\code{\link{make_ref2compressed}()},
\code{\link{runDiffSplice}()},
\code{\link{sortSamples}()},
\code{\link{spliceGR2junctionDF}()}
}
\concept{jam GRanges functions}
\concept{jam RNA-seq functions}
