% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{compressPolygonM}
\alias{compressPolygonM}
\title{Compress genome coordinates of a matrix of polygons}
\usage{
compressPolygonM(polyM, ref2c, minRatio = 3, verbose = FALSE, ...)
}
\description{
Compress genome coordinates of a matrix of polygons
}
\details{
This function takes a two-column numeric matrix of polygons
where the x coordinate is the genomic position, and y coordinate
is the coverage. It uses \code{ref2compressed$x} to convert coordinates
to compressed coordinates.

For regions that have been compresssed, it then compresses the
y-coordinate information to roughly one y value per integer in
compressed coordinate space, using the runmax across the window of
coverages compressed to this value.
}
