% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{defineDetectedTx}
\alias{defineDetectedTx}
\title{Define detected transcripts}
\usage{
defineDetectedTx(iMatrixTx = NULL, iMatrixTxGrp = NULL,
  iMatrixTxTPM = NULL, iMatrixTxTPMGrp = NULL, groups = NULL,
  tx2geneDF = NULL, cutoffTxPctMax = 10, cutoffTxExpr = 5,
  cutoffTxTPMExpr = 2, useMedian = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{iMatrixTx}{numeric matrix of read counts (or pseudocounts) with
transcript rows and sample columns.}

\item{iMatrixTxGrp}{numeric matrix of read counts averaged by sample
group. If this matrix is not provided, it will be calculated
from \code{iMatrixTx}
using \code{jamba::rowGroupMeans()} and the \code{groups} parameter.}

\item{iMatrixTxTPM}{numeric matrix of TPM values, with sample columns
and transcript rows. Note that if this parameter is not supplied,
the counts in \code{iMatrixTx} will be used to determine the percent
max isoform expression.}

\item{iMatrixTxTPMGrp}{numeric matrix of TPM values averaged by sample
group. If this matrix is not provided, it will be calculated
from \code{iMatrixTxTPM}
using \code{jamba::rowGroupMeans()} and the \code{groups} parameter.}

\item{groups}{vector of group labels, either as character vector
or factor. It should be named by \code{colnames(iMatrixTx)}.}

\item{tx2geneDF}{data.frame with colnames including
\code{c("transcript_id","gene_name")}, where the values in the
\code{"transcript_id"} column must match the \code{rownames(iMatrixTx)}.}

\item{cutoffTxPctMax}{numeric value scaled from 0 to 100
indicating the percentage of
the maximum isoform expression per gene, for an alternate isoform
to be considered for detection.}

\item{cutoffTxExpr}{numeric value indicating the minimum group mean
counts in \code{iMatrixTxGrp} for a transcript to be considered for
detection.}

\item{cutoffTxTPMExpr}{numeric value indicating the minimum group mean
TPM in \code{iMatrixTxTPMGrp} for a transcript to be considered for
detection.}

\item{useMedian}{logical indicating whether to use group median
values instead of group mean values.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\value{
List with the following elements:
\describe{
\item{txExprGrpTx}{Numeric matrix representing the expression
counts per transcript, grouped by \code{"gene_name"}.}
\item{txPctMaxGrpAll}{Numeric matrix representing the
percent expression of each transcript isoform per gene, as
compared to the highest expression of isoforms for that gene.}
\item{txExprGrpAll}{Numeric matrix of sample group counts,
exponentiated and rounded to integer values.}
\item{txTPMExprGrpAll}{Numeric matrix of sample group TPM values,
exponentiated and rounded to integer values.}
\item{txFilterM}{Numeric matrix indicating whether each isoform met
the criteria to be considered detected. The criteria must be
met in the same group for an isoform to be considered detected.}
\item{detectedTx}{Character vector of transcripts, as defined by
the \code{rownames(iMatrixTx)}.}
}
}
\description{
Define detected transcripts
}
\details{
This function aims to combine evidence from RNA-seq sequence
read counts (or pseudocounts from a kmer tool such as Salmon or
Kallisto), along with alternative TPM quantitation, to determine
the observed "detected" transcript space for a given experiment.

The criteria must be met in at least one sample group, but all
criteria must be met in the same sample group for an isoform to
be considered "detected".

In our experience the use of TPM values appears more robust and
is conceptually the best approach for comparing the relative
quantity of one transcript isoform to another. Our reasoning is
that TPM is intended to be roughly a molar quantity of transcript
molecules, independent of the transcript length, and the potential
for overlapping regions between isoforms. We also recommend the use
of a kmer quantitation method, such as Salmon or Kallisto, which
estimates isoform abundances not by specific read counts, but by
quantifying kmers unique to particular isoforms for a given gene.

In all cases, the thresholds for detection can be modified, however
from our experiences thus far the default values perform reasonably
well at identifying expressed isoforms, while filtering out isoforms
that we considered to be spuriously expressed.

There are three default requirements for a transcript to be considered
"detected".
\enumerate{
\item An isoform must be expressed at least 10% of the max isoform for
a given gene, using TPM values.
\item An isoform must have at least log2(32) pseudocounts to be considered
detected, based upon our view of Salmon pseudocount data using MA-plots.
\item An isoform must have at least log2(2) TPM units to be considered
detected, based upon our view of Salmon TPM values using MA-plots.
}

Each experiment is likely to be different in terms of total
sequenced reads, quality of read alignment or quantitation to the
transcriptome, etc. We suggest observing MA-plots for the counts and
TPM values, for the point at which the signal substantially increases
from baseline zero. We also plotted the TPM versus count per sample,
noted the point at which the two signals began to correlate. These
observations along with careful review of numerous gene model
transcript isoforms supported our selection of these criteria.

Lastly, the requirement for 10 percent of max isoform expression
was motivated by observing highly expressed genes, which sometimes had
alternative isoforms with extremely low abundance compared to the
most abundant isoform, but which was notably higher than the minimum
for detection. For example Gapdh expression above 100,000 pseudocounts,
may have an isoform with 120 pseudocounts. When we reviewed the sequence
coverage, we could find no compelling evidence to support the minor
isoform, and theorized that the pseudocounts arose from the stochastic
nature of rebalancing relative expression among isoforms.
}
\seealso{
Other jam RNA-seq functions: \code{\link{ale2violin}},
  \code{\link{codonUsage2df}},
  \code{\link{detectedTxInfo}}, \code{\link{dna2codon}},
  \code{\link{groups2contrasts}}, \code{\link{jamCai}},
  \code{\link{makeTx2geneFromGtf}},
  \code{\link{runDiffSplice}}, \code{\link{sortSamples}},
  \code{\link{tx2ale}}
}
\concept{jam RNA-seq functions}
