% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gtf.R
\name{describeGtfAttrNames}
\alias{describeGtfAttrNames}
\title{describe gtf/gff3 attribute names}
\usage{
describeGtfAttrNames(
  GTF,
  geneFeatureType = "gene",
  txFeatureType = c("transcript", "mRNA"),
  nrows = 10000,
  zcat_command = "zcat",
  verbose = FALSE,
  ...
)
}
\description{
describe gtf/gff3 attribute names
}
