% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{detectedTxInfo}
\alias{detectedTxInfo}
\title{Summarize detected transcript results}
\usage{
detectedTxInfo(detectedTxL, Gene = NULL, Tx = NULL, ...)
}
\arguments{
\item{detectedTxL}{list output from \code{defineDetectedTx()}}

\item{Gene}{optional character vector of one or more genes
of interest, used to find transcript_id entries in the
\code{detectedTxL} input data. If not supplied, \code{Tx} is expected.}

\item{Tx}{optional character vector used to subset summary data,
usually intended to keep rows in a specific order for a
given set of transcripts.}

\item{...}{additional arguments are ignored.}
}
\value{
list of \code{data.frame} objects, each containing one
summary table of data used to support whether each transcript
were called "detected."
}
\description{
Summarize detected transcript results
}
\details{
This function provides a simple summary of the results of
\code{defineDetectedTx()}, typically for a given gene of interest.
}
\seealso{
Other jam RNA-seq functions: \code{\link{ale2violin}},
  \code{\link{assignGRLexonNames}},
  \code{\link{closestExonToJunctions}},
  \code{\link{codonUsage2df}},
  \code{\link{combineGRcoverage}},
  \code{\link{compressPolygonM}},
  \code{\link{defineDetectedTx}}, \code{\link{dna2codon}},
  \code{\link{exoncov2polygon}},
  \code{\link{flattenExonsByGene}},
  \code{\link{groups2contrasts}}, \code{\link{jamCai}},
  \code{\link{makeTx2geneFromGtf}},
  \code{\link{make_ref2compressed}},
  \code{\link{runDiffSplice}}, \code{\link{sortSamples}},
  \code{\link{spliceGR2junctionDF}}, \code{\link{tx2ale}}
}
\concept{jam RNA-seq functions}
