% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{dna2codon}
\alias{dna2codon}
\title{Convert DNA to 3-base codons}
\usage{
dna2codon(x, ...)
}
\arguments{
\item{x}{character vector assumed to contain DNA or RNA nucleotides}

\item{...}{additional arguments are ignored.}
}
\value{
character vector where each element contains three characters,
e.g. \code{all(nchar(dna2codon(x)) == 3)}.
}
\description{
Convert DNA to 3-base codons
}
\details{
This function takes character string input, either as one large
string or vector of strings such as those read from a FASTA file,
and returns a vector of 3-base codons. The final codon must contain
all three character values otherwise it is removed.
}
\examples{
dna <- c("atgggattatag");
dna2codon(dna);

# example adding one extra base, which does not make a full codon
dnaext1 <- c("atgggattataga");
dna2codon(dnaext1);

}
\seealso{
Other jam RNA-seq functions: \code{\link{ale2violin}},
  \code{\link{assignGRLexonNames}},
  \code{\link{closestExonToJunctions}},
  \code{\link{codonUsage2df}},
  \code{\link{combineGRcoverage}},
  \code{\link{compressPolygonM}},
  \code{\link{defineDetectedTx}},
  \code{\link{detectedTxInfo}},
  \code{\link{exoncov2polygon}},
  \code{\link{flattenExonsByGene}},
  \code{\link{groups2contrasts}}, \code{\link{jamCai}},
  \code{\link{makeTx2geneFromGtf}},
  \code{\link{make_ref2compressed}},
  \code{\link{runDiffSplice}}, \code{\link{sortSamples}},
  \code{\link{spliceGR2junctionDF}}, \code{\link{tx2ale}}
}
\concept{jam RNA-seq functions}
