% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{exoncov2polygon}
\alias{exoncov2polygon}
\title{Convert exon coverage to polygons}
\usage{
exoncov2polygon(gr, covNames = NULL, baseline = 0, gapWidth = 250,
  doPlot = FALSE, verbose = FALSE, ...)
}
\description{
Convert exon coverage to polygons
}
