% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{findOverlapsGRL}
\alias{findOverlapsGRL}
\title{Find overlaps between two GRangesList objects}
\usage{
findOverlapsGRL(GRL1, GRL2, annoName1 = "name", annoName2 = "name",
  ...)
}
\arguments{
\item{GRL1}{GRangesList query}

\item{GRL2}{GRangesList subject}

\item{annoName1}{character value indicating either the colname of
\code{values(GRL1)} to use as the name, or if \code{"name"} then it uses
\code{names(GRL1)}.}

\item{annoName2}{character value indicating either the colname of
\code{values(GRL2)} to use as the name, or if \code{"name"} then it uses
\code{names(GRL2)}.}

\item{...}{additional arguments are passed to
\code{GenomicRanges::findOverlaps()}, useful for customizing the overlap
criteria.}
}
\value{
Hits object, or the natural output from
\code{GenomicRanges::findOverlaps()} dependent upon the \code{...} arguments.
}
\description{
Find overlaps between two GRangesList objects
}
\details{
This function implements \code{GenomicRanges::findOverlaps()} for the special
case of two GRangesList objects, restricting results to those including
the same GRangesList index in the subject and query.
}
\seealso{
Other jam GRanges functions: \code{\link{addGRLgaps}},
  \code{\link{addGRgaps}},
  \code{\link{annotateGRLfromGRL}},
  \code{\link{annotateGRfromGR}},
  \code{\link{assignGRLexonNames}},
  \code{\link{getFirstStrandedFromGRL}},
  \code{\link{sortGRL}}
}
\concept{jam GRanges functions}
