% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{geomean}
\alias{geomean}
\title{Classical geometric mean}
\usage{
geomean(x, na.rm = TRUE, naValue = NA, offset = 0, ...)
}
\arguments{
\item{x}{numeric vector containing only positive values}

\item{na.rm}{logical indicating whether to ignore \code{NA} values. Note that
\code{NA} values are removed prior to log-transformation, to avoid negative
numbers being dropped completely. To drop negative numbers, do so
prior to calling \code{geomean()}.}

\item{offset}{numeric value added to input \code{x} prior to log
transformation, intended only when values between 0 and 1 should be
retained. Note that the offset makes the result slightly different
than classical geometric mean.}

\item{...}{additional parameters are ignored.}
}
\value{
numeric value representing the geometric mean of input values
}
\description{
Classical geometric mean
}
\details{
This function calculates the classical geometric mean.

The classical geometric mean is defined as
the exponentiated mean of log-transformed values. Said another
way, it is the \code{n}th root of the product of \code{n} numeric values.
This formula is analogous to geometric distance. The formula
does not allow negative values, however, and if any value is
zero the result is also zero.
}
\examples{
x <- c(2, 4);
geomean(x);

x <- c(-2, 2, 4);
geomean(x);

x <- c(0, 4000, 200000);
geomean(x);

}
\seealso{
Other jam numeric functions: \code{\link{jamGeomean}}
}
\concept{jam numeric functions}
