% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getFirstStrandedFromGRL}
\alias{getFirstStrandedFromGRL}
\title{Get first stranded GRanges feature per GRangesList}
\usage{
getFirstStrandedFromGRL(grl, method = c("endoapply", "flank"),
  verbose = FALSE, ...)
}
\arguments{
\item{grl}{GRangesList}

\item{method}{character value in \code{c("endoapply", "flank")}
representing which method to use to define the first feature.
The \code{"endoapply"} method uses \code{S4Vectors::endoapply()} to
iterate each GRangesList element. The \code{"flank"} method is
intended to be equivalent but uses the \code{GenomicRanges::flank()}
function, which is vectorized.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
GRangesList containing only the first GRanges feature
in stranded order.
}
\description{
Get first stranded GRanges feature per GRangesList
}
\details{
This function returns the first feature per GRangesList,
relative to the strand of the GRangesList. It assumes each
GRangesList element has only one strand, as is true for exons
in one transcript.
}
\seealso{
Other jam GRanges functions: \code{\link{annotateGRLfromGRL}},
  \code{\link{annotateGRfromGR}},
  \code{\link{assignGRLexonNames}},
  \code{\link{findOverlapsGRL}}, \code{\link{sortGRL}}
}
\concept{jam GRanges functions}
