% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{getGRcoverageFromBw}
\alias{getGRcoverageFromBw}
\title{Get coverage for GRanges from bigWig files}
\usage{
getGRcoverageFromBw(gr, bwUrls, addGaps = FALSE,
  gap_feature_type = "gap", default_feature_type = "exon",
  feature_type_colname = "feature_type", verbose = FALSE, ...)
}
\arguments{
\item{gr}{GRanges object}

\item{bwUrls}{character vector of full file paths or web URLs
to bigWig files, suitable for use by \code{rtracklayer::import()}.}

\item{addGaps}{logical indicating whether gaps between GRanges
should be added to the query. Gaps are determined using
\code{getGRgaps()}.}

\item{feature_type_colname, gap_feature_type, default_feature_type}{When \code{addGaps=TRUE} a
new column named using \code{feature_type_colname} is added to \code{values(gr)},
whose value for gap regions is \code{gap_feature_type}. When
\code{feature_type_colname} is already present in \code{gr} it is not modified,
otherwise the column is created with value \code{default_feature_type}.
By default, this function adds a column \code{"feature_type"} with
value \code{"gap"}.}

\item{...}{additional arguments are ignored.}
}
\value{
DataFrame object, whose colnames are defined using
\code{makeNames(basename(bwUrls))}. Each column is type \code{NumericList},
which is a list of numeric coverage values.
}
\description{
Get coverage for GRanges from bigWig files
}
\details{
This function takes a GRanges object to define genomic regions,
for which coverage data is loaded for each \code{bwUrls} input file.
}
\seealso{
Other jam GRanges functions: \code{\link{addGRLgaps}},
  \code{\link{addGRgaps}},
  \code{\link{annotateGRLfromGRL}},
  \code{\link{annotateGRfromGR}},
  \code{\link{assignGRLexonNames}},
  \code{\link{combineGRcoverage}},
  \code{\link{exoncov2polygon}},
  \code{\link{findOverlapsGRL}},
  \code{\link{getFirstStrandedFromGRL}},
  \code{\link{getGRLgaps}}, \code{\link{getGRgaps}},
  \code{\link{grl2df}}, \code{\link{make_ref2compressed}},
  \code{\link{sortGRL}}

Other RNA-seq functions: \code{\link{prepareSashimi}}
}
\concept{RNA-seq functions}
\concept{jam GRanges functions}
