% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getGRgaps}
\alias{getGRgaps}
\title{Get gaps in GRanges}
\usage{
getGRgaps(gr, strandSpecific = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{gr}{GRanges object}

\item{strandSpecific}{logical indicating whether to convert strand
to \code{"*"} prior to determining gaps between features.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{getGRLgaps()}.}
}
\description{
Get gaps in GRanges
}
\details{
This function returns the gaps between GRanges regions, calculated
for each chromosome (using \code{GenomicRanges::seqnames(gr)}), and when \code{strandSpecific=TRUE}
it determines gaps in stranded fashion.
}
\seealso{
Other jam GRanges functions:
\code{\link[=addGRLgaps]{addGRLgaps()}},
\code{\link[=addGRgaps]{addGRgaps()}},
\code{\link[=annotateGRLfromGRL]{annotateGRLfromGRL()}},
\code{\link[=annotateGRfromGR]{annotateGRfromGR()}},
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=closestExonToJunctions]{closestExonToJunctions()}},
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=findOverlapsGRL]{findOverlapsGRL()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getFirstStrandedFromGRL]{getFirstStrandedFromGRL()}},
\code{\link[=getGRLgaps]{getGRLgaps()}},
\code{\link[=getGRcoverageFromBw]{getGRcoverageFromBw()}},
\code{\link[=grl2df]{grl2df()}},
\code{\link[=jam_isDisjoint]{jam_isDisjoint()}},
\code{\link[=make_ref2compressed]{make_ref2compressed()}},
\code{\link[=sortGRL]{sortGRL()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}},
\code{\link[=stackJunctions]{stackJunctions()}}
}
\concept{jam GRanges functions}
