% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getGRgaps}
\alias{getGRgaps}
\title{Get gaps in GRanges}
\usage{
getGRgaps(GR, strandSpecific = TRUE, keepValues = FALSE,
  trimEnds = TRUE, method = c("new", "old"), ...)
}
\description{
Get gaps in GRanges
}
