% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{getGRLgaps}
\alias{getGRLgaps}
\title{Get gaps in GRangesList objects}
\usage{
getGRLgaps(grl, strandSpecific = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{grl}{GRangesList object.}

\item{strandSpecific}{logical indicating whether to determine gaps
within strand.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
GRangesList object with gaps for each chromosome and strand
present in each GRanges entry. It does not return gap sequence
at the edges of GRanges regions to the chromosome ends.
}
\description{
Get gaps in GRangesList objects
}
\details{
This function returns gaps between GRanges regions in a GRangesList
object. When \code{strandSpecific=TRUE} is determines gaps per strand,
otherwise strands are converted to \code{"*"}. It will also determine
gaps within chromosome for each GRanges entry in GRangesList.
}
\seealso{
Other jam GRanges functions: 
\code{\link{addGRLgaps}()},
\code{\link{addGRgaps}()},
\code{\link{annotateGRLfromGRL}()},
\code{\link{annotateGRfromGR}()},
\code{\link{assignGRLexonNames}()},
\code{\link{closestExonToJunctions}()},
\code{\link{combineGRcoverage}()},
\code{\link{exoncov2polygon}()},
\code{\link{findOverlapsGRL}()},
\code{\link{getFirstStrandedFromGRL}()},
\code{\link{getGRcoverageFromBw}()},
\code{\link{getGRgaps}()},
\code{\link{grl2df}()},
\code{\link{make_ref2compressed}()},
\code{\link{sortGRL}()},
\code{\link{spliceGR2junctionDF}()},
\code{\link{stackJunctions}()}
}
\concept{jam GRanges functions}
