% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gtf.R
\name{getGtfAttrs}
\alias{getGtfAttrs}
\title{Grab GTF or GFF3 attributes into a data.frame}
\usage{
getGtfAttrs(
  gtfDF,
  useRows = NULL,
  attrNames,
  featureType = NULL,
  ignore.case = FALSE,
  matchMethod = c("once", "multi"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gtfDF}{\code{data.frame} with GTF or GFF3 formatted data}

\item{useRows}{\code{logical}, default NULL, rows in \code{gtfDF} to use for analysis,
or NULL to use all rows supplied in \code{gtfDF}.}

\item{attrNames}{\code{character} vector of attribute names, matching one
of two types of fields in the \code{gtfDF} data:
\enumerate{
\item Name/value pairs in column 9, typically used for miscellaneous
annotations. The formats are intended to represent accepted formats
according to the GTF and GFF3 specifications, for example:
\itemize{
\item GTF: \verb{attrName "value"}
\item GTF: \verb{attrName "value";}
\item GTF: \verb{attrName value}
\item GTF: \verb{attrName value;}
\item GFF3: \code{attrName=value}
\item GFF3: \code{attrName=value;}
\item GFF3: \code{attrName="value"}
\item GFF3: \code{attrName="value";}
}
\item Recognized column from within the GTF or GFF3 format:
\itemize{
\item \strong{chr} - column 1
\item \strong{start} - column 4
\item \strong{end} - column 5
\item \strong{strand} - column 7
\item \strong{range} - columns 1,4,5,7 concatenated into one field,
for example: "chr:start-end:strand" or "chr1:10000-11000:+"
}
}}

\item{featureType}{\code{character}, default NULL, with feature type used
as a prefix for coordinate/range attributes, and otherwise only
used in messaging when \code{verbose=TRUE}.}

\item{ignore.case}{\code{logical} default FALSE, passed to \code{grep()} and
\code{gsub()} to enable case-insensitive matching of attribute names.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} with colnames derived from the GTF or GFF3 data.
When there is no data recognized, it returns \code{NULL}, for example
when \code{useRows} is entirely \code{FALSE} or no \code{attrNames} are recognized
in the input \code{gtfDF}.
}
\description{
Grab GTF or GFF3 attributes into a data.frame
}
\details{
The input retrieves data for known attribute names, although when
an attribute name is not present, it is silently ignored.
However, to see available attribute names, try \code{describeGtfAttrNames()}.
}
\seealso{
Other jam gtf functions: 
\code{\link{describeGtfAttrNames}()},
\code{\link{makeTx2geneFromGtf}()},
\code{\link{readGtf}()}
}
\concept{jam gtf functions}
