% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gg.R
\name{grl2df}
\alias{grl2df}
\title{GRangesList to data.frame for ggplot2}
\usage{
grl2df(grl, keepGRvalues = TRUE, keepGRLvalues = FALSE,
  addGaps = TRUE, width = 0.6, widthV = c(exon = 0.6, intron = 0.02,
  gap = 0.02), width_colname = "feature_type", shape = c("rectangle",
  "junction"), baseline = NULL, scoreColname = "score",
  scoreMinimum = 200, scoreFactor = 1, scoreArcFactor = 0.5,
  doStackJunctions = TRUE, ref2c = NULL, verbose = FALSE, ...)
}
\description{
GRangesList to data.frame for ggplot2
}
