% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{import_juncs_from_bed}
\alias{import_juncs_from_bed}
\title{Import splice junction data from BED or SJ.out.tab file}
\usage{
import_juncs_from_bed(
  iBed,
  juncNames,
  sample_id,
  scale_factor,
  use_memoise = FALSE,
  memoise_junction_path = "junctions_memoise",
  gr = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{iBed}{path or URL to one BED file containing splice
junction data. The score is expected to be stored in the
name column (column 3), primarily because the score column
is sometimes restricted to maximum value 1000. However if
the name column cannot be converted to numeric without
creating NA values, then the score column will be used.}

\item{juncNames}{the name of the junction source file}

\item{sample_id}{character string representing the sample
identifier.}

\item{scale_factor}{numeric value used to adjust the raw
score, applied by multiplying the scale_factor by each score.}

\item{gr}{GRanges representing the overall range for which
junction data will be retrieved. Note that any junctions
that span this range, but do not start or end inside this
range, will be removed.}
}
\description{
Import splice junction data from BED or SJ.out.tab file
}
\details{
This function is intended to be called internally by
\code{prepareSashimi()}, and is provided primarily to enable
use of \code{memoise::memoise()} to cache results.

This function was refactored in version 0.0.69.900 to
handle either BED format, or \code{"SJ.out.tab"} junction
format as produced by STAR alignment. The method uses
\code{data.table::fread()} then if there are 9 columns,
it assumes the format is \code{"SJ.out.tab"}. Otherwise it
coerces the \code{data.frame} with \code{as(bed, "GRanges")}.

The BED or \code{"SJ.out.tab"} file can be gzipped, provided
\code{data.table()} is able to recognieze and import the
compression format.

Note that bigBed format still cannot be used since the
rtracklayer package does not support that format.

Also note that junctions with \code{score=0} are dropped at
this step, to prevent propagation of junctions with
zero counts.
}
\seealso{
Other jam data import functions: 
\code{\link{psl2df}()}
}
\concept{jam data import functions}
