% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{intercalate}
\alias{intercalate}
\title{Intercalate list values into a vector}
\usage{
intercalate(...)
}
\arguments{
\item{...}{parameters are expected as multiple vectors.}
}
\description{
Intercalate list values into a vector
}
\details{
This function takes a list of vectors, and intercalates them
by inserting alternating values from each vector to make one
combined vector. It is analogous to shuffling a deck of cards,
except that there could be more than two stacks of cards shuffled
together.
}
