% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{internal_junc_score}
\alias{internal_junc_score}
\title{maximum overlapping internal junction score}
\usage{
internal_junc_score(juncGR, scoreColname = "score",
  sampleColname = "sample_id", verbose = FALSE, ...)
}
\arguments{
\item{juncGR}{GRanges containing splice junctions, with numeric
column \code{scoreColname} representing the abundance of splice
junction-spanning sequence reads.}

\item{scoreColname, sampleColname}{colnames in \code{values(juncGR)}
which define the junction score, and the \code{"sample_id"}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
numeric vector named by \code{names(juncGR)} whose values
are the maximum score of internal overlapping junction ends.
}
\description{
maximum overlapping internal junction score
}
\details{
This function is intended for internal use, and calculates the
maximum score for junction GRanges for the special case where:
\itemize{
\item a junction range overlaps the start or end of another junction
\item the start or end of the other junction is internal, that is
it does not overlap the same start or end.
\item overlaps are constrained to the same \code{"sample_id"} stored
in \code{values(juncGR)$sample_id}.
}

The goal is to determine the highest score contained
inside each junction region, so that the junction arc height
can be defined higher in order to minimize junction arc
overlaps.
}
