% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{jam_isDisjoint}
\alias{jam_isDisjoint}
\title{Test whether GRanges are disjoint (non-overlapping)}
\usage{
jam_isDisjoint(x)
}
\arguments{
\item{x}{\code{GRanges} or \code{GRangesList} object}
}
\value{
\code{logical} indicating whether the input \code{x} contains
disjoint ranges. When input \code{x} is a \code{GRangesList} then this
function tests \code{GRanges} elements independently for overlaps,
and returns \code{TRUE} only if all \code{GRanges} elements are disjoint.
If any \code{GRanges} elements are not disjoint, this function
returns \code{FALSE}.
}
\description{
Test whether GRanges are disjoint (non-overlapping)
}
\details{
This function is a simple alternative to \code{GenomicRanges::isDisjoint()}
that is intended to produce identical output while being markedly
more efficient.
}
\seealso{
Other jam GRanges functions: 
\code{\link{addGRLgaps}()},
\code{\link{addGRgaps}()},
\code{\link{annotateGRLfromGRL}()},
\code{\link{annotateGRfromGR}()},
\code{\link{assignGRLexonNames}()},
\code{\link{closestExonToJunctions}()},
\code{\link{combineGRcoverage}()},
\code{\link{exoncov2polygon}()},
\code{\link{findOverlapsGRL}()},
\code{\link{getFirstStrandedFromGRL}()},
\code{\link{getGRLgaps}()},
\code{\link{getGRcoverageFromBw}()},
\code{\link{getGRgaps}()},
\code{\link{grl2df}()},
\code{\link{make_ref2compressed}()},
\code{\link{sortGRL}()},
\code{\link{spliceGR2junctionDF}()},
\code{\link{stackJunctions}()}
}
\concept{jam GRanges functions}
