% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{jam_isDisjoint}
\alias{jam_isDisjoint}
\title{Test whether GRanges are disjoint (non-overlapping)}
\usage{
jam_isDisjoint(x)
}
\arguments{
\item{x}{\code{GRanges} or \code{GRangesList} object}
}
\value{
\code{logical} vector indicating whether the input \code{x} contains
disjoint ranges. When input \code{x} is a \code{GRangesList} then the
vector represents each \code{GenomicRanges} object in the \code{GRangesList}
and should therefore be \code{length(x)}. When input \code{x} is
\code{GRanges} this function returns one value representing the
full \code{GRanges} object. The output for this function changed
in version \verb{0.0.64.900} to be consistent with
\code{GenomicRanges::isDisjoint()}.
}
\description{
Test whether GRanges are disjoint (non-overlapping)
}
\details{
This function is a simple alternative to \code{GenomicRanges::isDisjoint()}
that is intended to produce identical output while being markedly
more efficient. For large \code{GRangesList} objects the result
from \code{GenomicRanges::isDisjoint()} was returned in 100+ seconds,
and the result from \code{jam_isDisjoint()} was returned in 0.5 seconds.

Note that this function forces \code{ignore.strand=FALSE}, which means
all comparisons are strand-specific. To force non-stranded comparisons,
update the strand of the input object to \code{"*"}.
}
\seealso{
Other jam GRanges functions: 
\code{\link{addGRLgaps}()},
\code{\link{addGRgaps}()},
\code{\link{annotateGRLfromGRL}()},
\code{\link{annotateGRfromGR}()},
\code{\link{assignGRLexonNames}()},
\code{\link{closestExonToJunctions}()},
\code{\link{combineGRcoverage}()},
\code{\link{exoncov2polygon}()},
\code{\link{findOverlapsGRL}()},
\code{\link{flattenExonsBy}()},
\code{\link{getFirstStrandedFromGRL}()},
\code{\link{getGRLgaps}()},
\code{\link{getGRcoverageFromBw}()},
\code{\link{getGRgaps}()},
\code{\link{grl2df}()},
\code{\link{make_ref2compressed}()},
\code{\link{sortGRL}()},
\code{\link{spliceGR2junctionDF}()},
\code{\link{stackJunctions}()}
}
\concept{jam GRanges functions}
