% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny.R
\name{launchSashimiApp}
\alias{launchSashimiApp}
\title{Launch Sashimi R-shiny application}
\usage{
launchSashimiApp(..., options = list(width = 1200))
}
\arguments{
\item{...}{additional arguments are passed to \code{shiny::shinyApp()}.}
}
\description{
Launch Sashimi R-shiny application
}
\details{
This function launches the Sashimi visualization
R-shiny app. Input is taken from the global environment,
otherwise defaults are used from the Farris et al data
in the \code{farrisdata} R package at https://github.com/jmw86069/farrisdata.
}
\examples{
# Note: disabled for web page examples
# launchSashimiApp();

}
\seealso{
Other splicejam R-shiny functions: \code{\link{sashimiAppServer}}
}
\concept{splicejam R-shiny functions}
