% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny.R
\name{launchSashimiApp}
\alias{launchSashimiApp}
\title{Launch Sashimi R-shiny application}
\usage{
launchSashimiApp(...)
}
\arguments{
\item{...}{additional arguments are passed to \code{shiny::shinyApp()}.}
}
\description{
Launch Sashimi R-shiny application
}
\details{
This function launches the Sashimi visualization
R-shiny app.
}
\examples{
# Note: disabled for web page examples
# launchSashimiApp();

}
\seealso{
Other splicejam R-shiny functions: \code{\link{sashimiAppServer}}
}
\concept{splicejam R-shiny functions}
