% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{make_ref2compressed}
\alias{make_ref2compressed}
\title{Create a ref2compressed function to compress GR gaps}
\usage{
make_ref2compressed(
  gr,
  gapWidth = 200,
  keepValues = FALSE,
  upstream = 50000,
  upstreamGapWidth = gapWidth * 3,
  downstream = 50000,
  downstreamGapWidth = gapWidth * 3,
  nBreaks = 7,
  ignore.strand = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{GRanges} object containing regions not to compress. Regions
which are unstranded gaps are compressed to fixed width.}

\item{gapWidth}{integer value used for fixed gap width, or when
NULL the gap width is defined as 3 times the median feature width.}

\item{keepValues}{\code{logical} indicating whether to keep feature values
in the GRanges data, default FALSE.}

\item{upstream, downstream, upstreamGapWidth, downstreamGapWidth}{\code{integer} number of bases to extend upstream and downstream the
overall range of features, and the number of bases to compress
that extended range, respectively. Any coordinates upstream
by the 'upstream' amount will be compressed into 'upstreamGapWidth'
bases in visual space. This ratio represents a visual compression
multiplier. The compression is applied to upstream coordinate zero,
and to downstream coordinate 10 billion.
Defaults are 50000 for upstream and downstream, and three times
the gap width \code{3 * gapWidth}.}

\item{nBreaks}{\code{integer} number of x-axis coordinate breaks used
in ggplot labeling, default 7.}

\item{ignore.strand}{\code{logical} default TRUE, new as of 0.0.89.900,
will create compressed coordinates by ignoring strand-specificity.
Use FALSE for previous behavior, although it would be limited
to scenarios where two features on opposite strands should not
be used to produce one common set of "gaps" between the genome
region features. In most cases, typical splicejam functionality
is unaffected, but implies gaps should ignore strandedness.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
list with \code{trans_grc} which is class \code{"trans"} suitable
for use in ggplot2 functions; \code{transform} a function that converts
chromosome coordinates to compressed coordinates; \code{inverse} a function
that converts compressed coordinates to chromosome coordinates;
\code{scale_x_grc} a function used similar to \code{ggplot2::scale_x_continuous()}
during ggplot2 creation; \code{gr} a function that compresses coordinates
in a \code{GRanges} object; \code{grl} a function that compresses coordinates in
a \code{GRangesList} object. Attributes \code{"lookupCoordDF"} is a two-column
data.frame with chromosome coordinates and compressed coordinates,
which is used to create the other transformation functions via
\code{stats::approx()}; \code{"gapWidth"} the gap width used, since it can
be programmatically defined; \code{"gr"} the \code{GRanges} input data used
to train the transformation.
}
\description{
Create a ref2compressed function to compress GR gaps
}
\details{
This function takes a set of GRanges which are to be maintained
with fixed aspect ratio, and it defines a function to compress
coordinates of the gaps between GRanges features.
}
\examples{
suppressPackageStartupMessages(library(GenomicRanges));
suppressPackageStartupMessages(library(ggplot2));

data(test_exon_wide_gr);
# To plot a simple GRanges object
widedf <- grl2df(test_exon_wide_gr);
ggWide <- ggplot(widedf, aes(x=x, y=y, group=id, fill=feature_type)) +
   geom_polygon() +
   colorjam::theme_jam() +
   colorjam::scale_fill_jam() +
   xlab("chr1") +
   ggtitle("exons (introns as-is)")
print(ggWide);

# Now compress the introns keeping axis labels
ref2c <- make_ref2compressed(test_exon_wide_gr,
   nBreaks=10);
ggWide2 <- ggWide +
   scale_x_continuous(trans=ref2c$trans_grc) +
   xlab("chr1 (compressed introns)") +
   ggtitle("exons (compressed introns)")
print(ggWide2);

}
\seealso{
\code{grl2df()}, \code{test_junc_wide_gr}

Other jam GRanges functions:
\code{\link[=addGRLgaps]{addGRLgaps()}},
\code{\link[=addGRgaps]{addGRgaps()}},
\code{\link[=annotateGRLfromGRL]{annotateGRLfromGRL()}},
\code{\link[=annotateGRfromGR]{annotateGRfromGR()}},
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=closestExonToJunctions]{closestExonToJunctions()}},
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=findOverlapsGRL]{findOverlapsGRL()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getFirstStrandedFromGRL]{getFirstStrandedFromGRL()}},
\code{\link[=getGRLgaps]{getGRLgaps()}},
\code{\link[=getGRcoverageFromBw]{getGRcoverageFromBw()}},
\code{\link[=getGRgaps]{getGRgaps()}},
\code{\link[=grl2df]{grl2df()}},
\code{\link[=jam_isDisjoint]{jam_isDisjoint()}},
\code{\link[=sortGRL]{sortGRL()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}},
\code{\link[=stackJunctions]{stackJunctions()}}

Other splicejam core functions:
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=gene2gg]{gene2gg()}},
\code{\link[=grl2df]{grl2df()}},
\code{\link[=plotSashimi]{plotSashimi()}},
\code{\link[=prepareSashimi]{prepareSashimi()}},
\code{\link[=splicejamFigure]{splicejamFigure()}}

Other jam RNA-seq functions:
\code{\link[=assignGRLexonNames]{assignGRLexonNames()}},
\code{\link[=closestExonToJunctions]{closestExonToJunctions()}},
\code{\link[=combineGRcoverage]{combineGRcoverage()}},
\code{\link[=defineDetectedTx]{defineDetectedTx()}},
\code{\link[=detectedTxInfo]{detectedTxInfo()}},
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=flattenExonsBy]{flattenExonsBy()}},
\code{\link[=getGRcoverageFromBw]{getGRcoverageFromBw()}},
\code{\link[=groups2contrasts]{groups2contrasts()}},
\code{\link[=internal_junc_score]{internal_junc_score()}},
\code{\link[=makeTx2geneFromGtf]{makeTx2geneFromGtf()}},
\code{\link[=prepareSashimi]{prepareSashimi()}},
\code{\link[=runDiffSplice]{runDiffSplice()}},
\code{\link[=sortSamples]{sortSamples()}},
\code{\link[=spliceGR2junctionDF]{spliceGR2junctionDF()}}
}
\concept{jam GRanges functions}
\concept{jam RNA-seq functions}
\concept{splicejam core functions}
