% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{make_ref2compressed}
\alias{make_ref2compressed}
\title{Create a ref2compressed function to compress GR gaps}
\usage{
make_ref2compressed(gr, gapWidth = 200, keepValues = FALSE,
  upstream = 50000, upstreamGapWidth = gapWidth * 3,
  downstream = 50000, downstreamGapWidth = gapWidth * 3,
  refGR = NULL, refGRnames = NULL, verbose = FALSE, ...)
}
\description{
Create a ref2compressed function to compress GR gaps
}
\details{
This function takes a set of GRanges which are to be maintained
with fixed aspect ratio, and it defines a function to compress
coordinates of the gaps between GRanges features.
}
