% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{makeTx2geneFromGtf}
\alias{makeTx2geneFromGtf}
\title{Make tx2gene data.frame from a GTF file}
\usage{
makeTx2geneFromGtf(GTF, geneAttrNames = c("gene_id", "gene_name",
  "gene_type"), txAttrNames = c("transcript_id", "transcript_type"),
  geneFeatureType = "gene", txFeatureType = c("transcript", "mRNA"),
  nrows = -1L, verbose = FALSE, ...)
}
\arguments{
\item{GTF}{character file name sent to \code{data.table::fread()}. When the
file ends with ".gz", gzcat is used to gunzip the file during the
import step. (TODO: verify that handling of gz files is portable across
architectures, or gracefully exit as needed.)}

\item{geneAttrNames}{character vector of recognized attribute names
as they appear in column 9 of the GTF file, for gene rows.}

\item{txAttrNames}{character vector of recognized attribute names
as they appear in column 9 of the GTF file, for transcript rows.}

\item{geneFeatureType}{character value to match column 3 of the GTF
file, used to define gene rows, by default "gene".}

\item{txFeatureType}{character value to match column 3 of the GTF
file, used to define gene rows, by default "transcript". In some
GTF files, "mRNA" is used, so either is accepted by default.}

\item{nrows}{integer number of rows to read from the GTF file, by default
-1 means all rows are imported. This parameter is useful to check the
results of a large GTF file using only a subset portion of the file.}

\item{verbose}{logical whether to print verbose output during processing.}
}
\value{
\code{data.frame} with colnames indicated by the values in
\code{geneAttrNames} and \code{txAttrNames}.
}
\description{
Make tx2gene data.frame from a GTF file
}
\details{
Create a transcript-to-gene data.frame from a GTF file, which is required
by a number of transcriptome analysis methods such as those in
the DEXseq package, and the limma package functions such as
\code{diffSplice()}.

This function also only uses \code{data.table::fread()} and does not
import the full GTF file using something like Bioconductor
\code{GenomicFeatures}, simply because the data.table method is markedly
faster when importing only the transcript-to-gene relationship. Also, this
method allows the import of more annotations than are supported by the
typical Bioconductor \code{rtracklayer::import()} for GTF data.

This function is intended to help keep all transcript data consistent by
using the same GTF file that is also used by other analysis tools, whether
those tools be based in R or more likely, outside R in a terminal
environment.

For example, the GTF file could be used:

\itemize{
\item{to run STAR sequence alignment
then \code{Rsubread::featureCounts()} to generate a matrix of read
counts per gene, transcript, or exon; or}
\item{to generate a transcript
FASTA sequence file then run a kmer quantitation tool such as
Salmon or Kallisto, then using \code{tximport::tximport()} to import
results into R for downstream processing.}
}
}
\seealso{
Other jam RNA-seq functions: \code{\link{ale2violin}},
  \code{\link{assignGRLexonNames}},
  \code{\link{codonUsage2df}},
  \code{\link{combineGRcoverage}},
  \code{\link{compressPolygonM}},
  \code{\link{defineDetectedTx}},
  \code{\link{detectedTxInfo}}, \code{\link{dna2codon}},
  \code{\link{flattenExonsByGene}},
  \code{\link{groups2contrasts}}, \code{\link{jamCai}},
  \code{\link{make_ref2compressed}},
  \code{\link{runDiffSplice}}, \code{\link{sortSamples}},
  \code{\link{tx2ale}}
}
\concept{jam RNA-seq functions}
