% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gg.R
\name{plotSashimi}
\alias{plotSashimi}
\title{Jam Sashimi plot}
\usage{
plotSashimi(sashimi, show = c("coverage", "junction", "junctionLabels"),
  coord_method = c("scale", "coord", "none"), exonsGrl = NULL,
  junc_color = alpha2col("goldenrod3", 0.7),
  junc_fill = alpha2col("goldenrod1", 0.4), fill_scheme = c("exon",
  "sample_id"), use_jam_themes = TRUE, apply_facet = TRUE,
  ref2c = NULL, verbose = FALSE, ...)
}
\arguments{
\item{sashimi}{Sashimi data prepared by \code{prepareSashimi()} which
is a \code{list} with \code{covDF} coverage data in data.frame format,
\code{juncDF} junction data in data.frame format,
\code{juncLabelDF} junction label coordinates in data.frame format,
\code{exonLabelDF} exon label coordinates per coverage polygon in
data.frame format,
\code{ref2c} list output from \code{make_ref2compressed()} to
transform genomic coordinates.}

\item{show}{character vector of Sashimi plot features to include:
\code{"coverage"} sequence read coverage data;
\code{"junction"} splice junction read data.}

\item{coord_method}{character value indicating the type of
coordinate scaling to use:
\code{"scale"} uses \code{ggplot2::scale_x_continuous()};
\code{"coord"} uses \code{ggplot2::coord_trans()};
\code{"none"} does not compress genomic coordinates.}

\item{exonsGrl}{GRangesList object with one or more gene or
transcript exon models, where exons are disjoint (not
overlapping.)}

\item{junc_color, junc_fill}{character string with valid R color,
used for junction outline, and fill, for the junction arc
polygon. Alpha transparency is recommended for \code{junc_fill}
so overlapping junction arcs are visible.}

\item{fill_scheme}{character string for how the exon coverages
will be color-filled: \code{"exon"} will define colors for each
distinct exon, using the GRanges names from \code{flatExonsByGene};
\code{"sample_id"} to color all exons the same by sample_id.}

\item{ref2c}{optional output from \code{make_ref2compressed()} used to
compress axis coordinates during junction arc calculations.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are sent to \code{grl2df()}.}
}
\description{
Jam Sashimi plot
}
\details{
This function uses Sashimi data prepared by \code{prepareSashimi()}
and creates a ggplot graphical object ready for visualization.
As a result, this function provides several arguments to
customize the visualization.
}
\examples{
suppressPackageStartupMessages(library(GenomicRanges));
data(test_exon_gr);
data(test_junc_gr);
data(test_cov_gr);
filesDF <- data.frame(url="sample_A",
   type="coverage_gr",
   sample_id="sample_A");
sh1 <- prepareSashimi(GRangesList(TestGene1=test_exon_gr),
   filesDF=filesDF,
   gene="TestGene1",
   covGR=test_cov_gr,
   juncGR=test_junc_gr);
plotSashimi(sh1);

}
\seealso{
Other jam plot functions: \code{\link{bgaPlotly3d}},
  \code{\link{factor2label}}, \code{\link{gene2gg}},
  \code{\link{grl2df}}, \code{\link{prepareSashimi}},
  \code{\link{stackJunctions}}

Other jam ggplot2 functions: \code{\link{gene2gg}}

Other splicejam core functions: \code{\link{exoncov2polygon}},
  \code{\link{gene2gg}}, \code{\link{grl2df}},
  \code{\link{make_ref2compressed}},
  \code{\link{prepareSashimi}}
}
\concept{jam ggplot2 functions}
\concept{jam plot functions}
\concept{splicejam core functions}
