% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{psl2df}
\alias{psl2df}
\title{Convert PSL alignment to data.frame}
\usage{
psl2df(psl, ...)
}
\arguments{
\item{psl}{file or other connection compatible with \code{base::readLines()}
of data in PSL alignment format.}

\item{...}{additional arguments are ignored.}
}
\description{
Convert PSL alignment to data.frame
}
\details{
This function takes PSL alignment format, as produced by BLAT,
and converts to data.frame.
}
