% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-import.R
\name{psl2df}
\alias{psl2df}
\title{Convert PSL alignment to data.frame}
\usage{
psl2df(psl, verbose = FALSE, ...)
}
\arguments{
\item{psl}{file or other connection compatible with \code{base::readLines()}
of data in PSL alignment format.}

\item{...}{additional arguments are ignored.}
}
\description{
Convert PSL alignment to data.frame
}
\details{
This function takes PSL alignment format, as produced by BLAT,
and converts to data.frame. The driving reason for this function
is that current methods to convert PSL to other formats lose
some useful information, for example conversion to BED12 format
loses the query alignment coordinates, in favor of storing
only the reference coordinates.

To supply text as input use \code{base::textConnection()} to wrap
a text connection around the input text.
}
\examples{

psls <- c(
   "2252\\t0\\t0\\t0\\t0\\t0\\t5\\t56285\\t+\\tQuery_Sequence\\t2252\\t0\\t2252\\tReference_Sequence\\t99095\\t1992\\t60529\\t6\\t227,86,71,79,77,1712,\\t0,227,313,384,463,540,\\t1992,6931,11020,39871,45861,58817,",
   "664\\t0\\t0\\t0\\t1\\t3\\t1\\t1\\t+\\tQuery_Sequence\\t2252\\t1291\\t1958\\tReference_Sequence\\t99095\\t58842\\t59507\\t2\\t657,7,\\t1291,1951,\\t58842,59500,"
   );
psl2df(textConnection(psls), verbose=TRUE)

}
\concept{jam data import functions}
