% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny-constants.R
\name{sashimiAppConstants}
\alias{sashimiAppConstants}
\title{SALSA Shiny app constants}
\usage{
sashimiAppConstants(...)
}
\description{
SALSA Shiny app constants
}
\details{
This function is intended to define constant values used in
the creation of the SALSA shiny UI.

This function is intended to be called only from within
an R-shiny app, since it defines several variables in the
parent (global) environment.
}
\concept{SALSA Shiny functions}
