% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny-constants.R
\name{sashimiAppConstants}
\alias{sashimiAppConstants}
\title{SALSA Shiny app constants}
\usage{
sashimiAppConstants(...)
}
\description{
SALSA Shiny app constants
}
\details{
This function is intended to define constant values used in
the creation of the SALSA shiny UI.

This function is intended to be called only from within
an R-shiny app, since it defines several variables in the
parent (global) environment.

Specifically, this function will attempt to re-use several
R data objects by name if they exist in any parent environment.
If they do not exist, defaults will be used based upon the
Farris et all data from the \code{farrisdata} package.
\itemize{
\item filesDF
\item color_sub
\item txdb (used to derive exonsByTx, cdsByTx if either does not exist)
\item tx2geneDF
\item gtf (used if tx2geneDF or exonsByTx are not available)
\item exonsByTx
\item cdsByTx
\item detectedTx (derived either from farrisdata::farrisTxSE
\code{"TxDetectedByTPM"} or using all entries in \code{tx2geneDF$transcript_id})
\item detectedGenes (inferred using tx2geneDF and detectedTx)
\item flatExonsByGene (defined using exonsByTx, cdsByTx, detectedTx,
tx2geneDF)
\item flatExonsByTx (defined using exonsByTx, cdsByTx, detectedTx,
tx2geneDF)
}
}
\concept{SALSA Shiny functions}
