% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny-server.R
\name{sashimiAppServer}
\alias{sashimiAppServer}
\title{Sashimi Shiny app server}
\usage{
sashimiAppServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{stssion}{provided by shiny}
}
\description{
Sashimi Shiny app sserver
}
\seealso{
Other splicejam R-shiny functions: \code{\link{launchSashimiApp}}
}
\concept{splicejam R-shiny functions}
