% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny-ui.R
\name{sashimiAppUI}
\alias{sashimiAppUI}
\title{Sashimi Shiny app UI}
\usage{
sashimiAppUI(...)
}
\arguments{
\item{...}{additional arguments are ignored.}
}
\description{
Sashimi Shiny app UI
}
\details{
This function contains the UI for the R-shiny
Splicejam Sashimi viewer.

The R-shiny app is started by \code{launchSashimiApp()}, which
calls \code{shiny::shinyApp()}, using arguments \code{server}, \code{ui},
\code{onStart}, and \code{options}. This function fulfills the
argument \code{ui}.
}
\seealso{
Other splicejam R-shiny functions: 
\code{\link{launchSashimiApp}()},
\code{\link{sashimiAppConstants}()},
\code{\link{sashimiAppServer}()},
\code{\link{sashimiDataConstants}()}
}
\concept{splicejam R-shiny functions}
