% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-shiny-ui.R
\name{sashimiAppUI}
\alias{sashimiAppUI}
\title{Sashimi Shiny app UI}
\usage{
sashimiAppUI(...)
}
\arguments{
\item{...}{additional arguments are ignored.}
}
\description{
Sashimi Shiny app UI
}
\concept{Sashimi Shiny functions}
