% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-tx.R
\name{shrinkMatrix}
\alias{shrinkMatrix}
\title{Shrink numeric matrix by groups of rows}
\usage{
shrinkMatrix(x, groupBy, shrinkFunc = function(x) {    
  .Internal(mean(x)) }, returnClass = c("data.frame", "matrix"),
  verbose = FALSE, ...)
}
\arguments{
\item{x}{numeric matrix}

\item{groupBy}{vector of group labels, whose length equals \code{nrow(x)}.
These values will become rownames in the output data.}

\item{returnClass}{character string indicating the return data type,
\code{"data.frame"} returns a \code{data.frame} whose first column contains
entries from \code{groups}; \code{"matrix"} returns a numeric matrix whose
rownames are entries from \code{groups}.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\description{
Shrink numeric matrix by groups of rows
}
\details{
This function is mainly a wrapper around the very efficient
functions in the \code{data.table} package, with the ability to
provide a custom function to shrink row values.
}
\concept{jam matrix functions}
