% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{simplifyXY}
\alias{simplifyXY}
\title{Simplify XY coordinates to minimal line segments}
\usage{
simplifyXY(xy, minN = 3, restrictDegrees = NULL, ...)
}
\arguments{
\item{xy}{numeric matrix with two columns representing x,y coordinates.}

\item{minN}{integer value to define the minimal number of repeated
values before the compression is used. By definition, three consecutive
points must have the same slope in order for compression to be
effective, otherwise the original coordinates will be returned.}

\item{restrictDegrees}{numeric vector of degrees to restrict the
simplification. For exmample \code{restrictDegrees=c(0,180)} will only
simplify horizontal lines.}

\item{...}{additional arguments are ignored.}
}
\description{
Simplify XY coordinates to minimal line segments
}
\details{
This function takes a numeric matrix of x,y coordinates
and returns the minimal matrix of x,y coordinates that represents
the same line segments. It is intended in cases where there
is a long repeated line segment that could be represented by
far fewer points.
}
\examples{
xy <- cbind(
   x=c(1,1:15,1),
   y=c(0,0,0,0,1,2,3,4,5,5,5,5,6,0,0));
par("mfrow"=c(1,1));
plot(xy, cex=2);
points(simplifyXY(xy), pch=20, col="orange", cex=3);
points(simplifyXY(xy, restrictDegrees=c(0,180)), pch=17, col="purple");
legend("topleft", pch=c(1, 20, 17),# box.col="black", border="black",
   pt.cex=c(2, 3, 1),
   col=c("black", "orange", "purple"),
   legend=c("all points", "simplified points", "only simplified horizontal"));

}
