% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sjenvtest.R
\docType{data}
\name{sjenvtest}
\alias{sjenvtest}
\title{Splicejam Environment Test Data}
\format{
\code{environment} with components suitable as
input to \code{splicejamFigure()} and other functions
which accept environment as input, typically
using argument 'sjenv'.
\describe{
\item{tx2geneDF}{\code{data.frame} with colnames including
the minimum required: 'gene_name', 'transcript_id'.}
\item{detectedGenes}{\code{character} vector of gene symbols.}
\item{detectedTx}{\code{character} vector of 'transcript_id'.}
\item{flatExonsByGene}{\code{GRangesList} named by 'gene_name'}
\item{flatExonsByTx}{\code{GRangesList} named by 'transcript_id'}
\item{txdb}{\code{TxDb} object used to create various exon
\code{GRanges} and \code{GRangesList} intermediate objects.
Not essential for processing, but often useful.}
\item{filesDF}{\code{data.frame} with colnames including
the minimum required: 'sample_id', 'url', 'type'}
}
}
\usage{
sjenvtest
}
\description{
Splicejam Environment Test Data, minimal subset of
Farris et al bulk RNA-seq data from mouse hippocampus.
It contains only 'CellType' values CA1 and CA2, with
both 'Compartment' values CB (cell body) and
DE (dendrites).
}
\seealso{
Other splicejam data:
\code{\link{test_cov_gr}},
\code{\link{test_cov_wide_gr}},
\code{\link{test_exon_gr}},
\code{\link{test_exon_wide_gr}},
\code{\link{test_junc_gr}},
\code{\link{test_junc_wide_gr}}
}
\concept{splicejam data}
\keyword{datasets}
