% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-design.R
\name{sortSamples}
\alias{sortSamples}
\title{Sort biological sample labels for experimental design}
\usage{
sortSamples(x, controlTerms = c("WT", "(^|[-_ ])NT($|[-_ ]|[0-9])",
  "ETOH", "control|ctrl|ctl", "ntc", "Vehicle|veh", "none|empty|blank",
  "scramble", "ttx", "PBS"), sortFunc = jamba::mixedSort,
  preControlTerms = NULL, postControlTerms = NULL,
  keepFactorsAsIs = TRUE, ...)
}
\arguments{
\item{x}{character vector or factor}

\item{controlTerms}{vector of regular expression patterns used to
determine control terms, where the patterns are matched and
returned in order.}

\item{preControlTerms}{vector or NULL, optional control
terms or regular expressions to use before the \code{controlTerms}
above. This argument is used as a convenient prefix to the
default terms.}

\item{postControlTerms}{vector or NULL, optional control
terms or regular expressions to use after the \code{controlTerms}
above. This argument is used as a convenient suffix to the
default terms.}

\item{keepFactorsAsIs}{logical indicating whether to maintain
factor level order, if \code{x} is supplied as a factor. If
\code{keepFactorsAsIs==TRUE} then only \code{sort(x)} is returned.}

\item{...}{additional arguments are ignored.}
}
\value{
character vector ordered such that control terms are
preferentially first before non-control terms.
}
\description{
Sort biological sample labels for experimental design
}
\details{
This function sorts a vector of sample labels using typical
heuristics that order typical control groups terms before
test groups. For example, \code{"Vehicle"} would be returned
before \code{"Treatment"} since \code{"Vehicle"} is a recognized control
term.

It also employs \code{jamba::mixedSort()} for
proper alphanumeric sorting, for example so \code{"Time_5hr"} would
be sorted before \code{"Time_12hr"}.
}
\examples{
# the defaults perform well for clear descriptors
sortSamples(c("Trt_12h", "Trt_9h", "Trt_1h", "Trt_9h", "Vehicle"));

# custom terms can be added before the usual control terms
sortSamples(c("Trt_12h", "Trt_9h", "Trt_1h", "Trt_9h", "Fixated", "Vehicle"),
   preControlTerms="fixate");

# custom terms can be added after the usual control terms
sortSamples(c("Trt_12h", "Trt_9h", "Trt_1h", "Trt_9h", "Fixated", "Vehicle"),
   postControlTerms="fixate");

}
\seealso{
Other jam string functions: \code{\link{escapeWhitespaceRegexp}},
  \code{\link{strsplitOrdered}}

Other jam RNA-seq functions: \code{\link{ale2violin}},
  \code{\link{assignGRLexonNames}},
  \code{\link{closestExonToJunctions}},
  \code{\link{codonUsage2df}},
  \code{\link{combineGRcoverage}},
  \code{\link{compressPolygonM}},
  \code{\link{defineDetectedTx}},
  \code{\link{detectedTxInfo}}, \code{\link{dna2codon}},
  \code{\link{exoncov2polygon}},
  \code{\link{flattenExonsBy}},
  \code{\link{groups2contrasts}}, \code{\link{jamCai}},
  \code{\link{makeTx2geneFromGtf}},
  \code{\link{make_ref2compressed}},
  \code{\link{runDiffSplice}},
  \code{\link{spliceGR2junctionDF}}, \code{\link{tx2ale}}
}
\concept{jam RNA-seq functions}
\concept{jam string functions}
