% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-plot.R
\name{spliceGR2junctionDF}
\alias{spliceGR2junctionDF}
\title{Splice junction data.frame summary}
\usage{
spliceGR2junctionDF(spliceGRgene, exonsGR, spliceBuffer = 3,
  spliceGRname = "score", geneExonSep = "(:|_exon)",
  useOnlyValidEntries = FALSE, renameTooFar = TRUE,
  scoreColname = "score", sampleColname = "sample_id",
  flipNegativeStrand = TRUE, returnGRanges = FALSE, verbose = FALSE,
  ...)
}
\description{
Splice junction data.frame summary
}
