% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-geoms.R
\name{splicejam-extensions}
\alias{splicejam-extensions}
\alias{StatDiagonalWideArc}
\title{splicejam extensions to ggforce}
\description{
ggforce makes heavy use of the \code{ggproto} class system to extend the
functionality of ggplot2. In general the actual classes should be of little
interest to users as the standard ggplot2 api of using geom_* and stat_*
functions for building up the plot is encouraged.
}
\seealso{
Other jam ggplot2 functions:
\code{\link[=gene2gg]{gene2gg()}},
\code{\link[=geom_diagonal_wide_arc]{geom_diagonal_wide_arc()}},
\code{\link[=plotSashimi]{plotSashimi()}},
\code{\link[=stat_unpack_polygon]{stat_unpack_polygon()}},
\code{\link[=to_basic.GeomShape]{to_basic.GeomShape()}}
}
\concept{jam ggplot2 functions}
