% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-figure.R
\name{splicejamFigure}
\alias{splicejamFigure}
\title{Make Splicejam Sashimi Figure}
\usage{
splicejamFigure(
  sjenv = NULL,
  gene = "Myom1",
  gene_sd = NULL,
  filesDF = NULL,
  sample_id = NULL,
  use_exon_range = NULL,
  display_coords = NULL,
  facet_scales = "free_y",
  minJunctionScore = 10,
  scoreArcFactor = 0.6,
  use_ylim = NULL,
  base_size = 18,
  exonLabelSize = 14,
  label_junctions = TRUE,
  junction_alpha = 0.8,
  layout_ncol = 1,
  gene_panel_height = 1,
  do_plot = TRUE,
  use_memoise = FALSE,
  panel_method = c("patchwork", "cowplot", "plotly"),
  show_progress = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sjenv}{\code{environment} containing data produced by
\code{sashimiDataConstants()}, specifically:
\itemize{
\item 'flatExonsByGene'
\item 'flatExonsByTx'
\item 'detectedTx'
\item 'detectedGenes'
\item tx2geneDF': \code{data.frame} with columns: 'gene_name',
'transcript_id'.
}}

\item{gene}{\code{character} string with gene(s) to include in the
flattened exon model.}

\item{gene_sd}{\code{list} optional sashimi data output from
\code{prepareSashimi()}.}

\item{filesDF}{\code{data.frame} with columns 'url', 'sample_id', 'type'.
See \code{prepareSashimi()} for more information.}

\item{sample_id}{\code{character} default NULL, optionally used to
subset data to show only data for samples 'sample_id'.}

\item{use_exon_range}{\code{character} vector with first and last exons
to include in the genome coordinate range. Typically in the
form: 'genesymbol_exon1', 'genesymbol_exon8b'}

\item{display_coords}{\code{numeric} default NULL, with optional genome
coordinates to use for the x-axis range.}

\item{facet_scales}{\code{character} string passed to
\code{ggplot2::facet_wrap()} argument 'scales'. Use 'fixed' to share
the same y-axis range in each panel.}

\item{minJunctionScore}{\code{numeric} default 10, passed to
\code{prepareSashimi()} to require junctions with at least this score
in order to be displayed. It is included here as a common
argument to customize via \code{prepareSashimi()}.}

\item{scoreArcFactor}{\code{numeric} default 0.6 with default arc height
added to the mean ribbon height, relative to its width.
Default 0.6 ensures the ribbon rises at least 60\% its width.
Note argument 'scoreArcMinimum=100' can be adjusted through
'...', and is not modified otherwise.}

\item{use_ylim}{\code{list} of \code{numeric} y-axis limits, applied to
each 'sample_id' panel. It uses \code{ggh4x::scale_y_facet()} and
is currently experimental, working out the enquo mechanics.}

\item{base_size}{\code{numeric} with base font size, default 18.}

\item{exonLabelSize}{\code{numeric} with exon label size, default 14.}

\item{label_junctions}{\code{logical} whether to label junctions with
count/score, default TRUE.}

\item{junction_alpha}{\code{numeric} default 0.8, default alpha for
junction ribbon arcs. Higher values are more opaque, and less
transparent.}

\item{layout_ncol}{\code{integer} number of columns for output, default 1.}

\item{gene_panel_height}{\code{numeric} relative panel height for
the gene/transcript/exon model, relative to each sashimi panel
with relative height 1. Default is 1.}

\item{do_plot}{\code{logical} default TRUE, whether to render the plot.}

\item{use_memoise}{\code{logical} default FALSE, whether to enable
data cache methods with memoise.}

\item{panel_method}{\code{character} method to assemble plot panels.
\itemize{
\item 'patchwork' (default) uses patchwork, and tends to have better
handling of whitespace
\item 'cowplot' uses cowplot, but tends to have much more whitespace.
\item 'plotly' uses plotly to create an interactive figure.
}}

\item{show_progress}{\code{logical} default TRUE, shows a progress bar
using R terminal or Shiny, depending upon whether it is running
inside Shiny.
\itemize{
\item \strong{Note:} The progress is updated (somewhere) but not displayed
unless the user (you) also enable it like this:
\code{progressr::handlers(global=TRUE)}
\item The exception is when called from within \code{sashimiAppServer()},
which sets \code{progressr::withShinyProgress()} to enable it.
}}

\item{verbose}{\code{logical} whether to print verbose output, default TRUE.
\itemize{
\item \code{verbose == TRUE}: prints major steps and elapsed time (default)
\item \code{verbose == 2}: as above, adding \code{prepareSashimi()} major steps
\item \code{verbose == 3}: as above, adding \code{plotSashimi()} and \code{use_ylim}
in detail. Probably too verbose at that point.
}}

\item{...}{additional arguments are passed to relevant functions:
\code{prepareSashimi()}, \code{gene2gg()}, \code{plotSashimi()}.
Some common examples:
\itemize{
\item show: passed to \code{plotSashimi()} to customize which features
are displayed: 'coverage', 'junctions', 'junctionLabels'
\item fill_scheme: passed to \code{plotSashimi()} to define coloring by
'sample_id' (default) or 'exon'. When 'exon' the junctions are
colored using 'junc_color' and 'junc_fill'.
}}
}
\value{
\code{list} with data elements:
\itemize{
\item cp_sashimi: the ggplot sashimi object
\item cp_sashimi_list: \code{list} of ggplot sashimi objects, per panel
\item cp_gene: the ggplot gene/transcript/exon object
\item cp_genes: \code{list} of ggplot gene/transcript/exon objects
\item gene_sd: the sashimi data from \code{prepareSashimi()}
\item layout_ncol: the layout ncol as used
\item use_ylim: the \code{list} of ylim as used
\item cp: the final assembled ggplot figure
}
}
\description{
Make Splicejam Sashimi Figure, with optional columns, gene model,
subset region by coordinates or exon range.
}
\details{
This function is intended to mimic the functional steps
performed by \code{launchSashimiApp()} which are aimed to provide
a user-friendly interface to the many internal details.

Notable features:
\itemize{
\item 'use_exon_range': new argument allows zooming into a genome
coordinate range using exon names.
\item gene model is shown and aligned by default, using cowplot.
\item 'layout_ncol': allows multiple columns with splicejam and
gene/transcript/exon model.
}
\subsection{Todo:}{
\itemize{
\item Consider providing interactive output via \code{ggplotly()}.
\item Expand the approach to permit multiple genes.
\item Apply memoise to plot data, as in \code{launchSashimiApp()}
\item Check when supplied with display_coords, no use_exon_range,
it should (in theory) find gene(s) in that region, use
them to assemble flattened exons, then generate figure with
one or more genes.
}
}
}
\examples{
# By default, it loads farrisdata
sjenv <- sashimiDataConstants()

sjenv$filesDF <- farrisdata::farris_sashimi_files_df
sjfig <- splicejamFigure(sjenv=sjenv, gene="Gria1", sample_id=c("CA2_CB", "CA2_DE"))

sjfig <- splicejamFigure(sjenv=sjenv, gene="Gria1", sample_id=c("CA2_CB", "CA2_DE"),
   use_exon_range=c("Gria1_exon13", "Gria1_exon16"), scoreArcFactor=0.4)

sjfig2 <- splicejamFigure(sjenv=sjenv, gene="Gria1", sample_id=c("CA2_CB", "CA2_DE"),
   layout_ncol=2,
   use_exon_range=c("Gria1_exon13", "Gria1_exon16"), scoreArcFactor=0.4)

sjfig3 <- splicejamFigure(sjenv=sjenv, gene="Gria1", sample_id=c("CA2_CB", "CA2_DE", "CA1_CB"),
   layout_ncol=2, use_memoise=TRUE,
   use_exon_range=c("Gria1_exon13", "Gria1_exon16"), scoreArcFactor=0.4)

}
\seealso{
Other splicejam core functions:
\code{\link[=exoncov2polygon]{exoncov2polygon()}},
\code{\link[=gene2gg]{gene2gg()}},
\code{\link[=grl2df]{grl2df()}},
\code{\link[=make_ref2compressed]{make_ref2compressed()}},
\code{\link[=plotSashimi]{plotSashimi()}},
\code{\link[=prepareSashimi]{prepareSashimi()}}
}
\concept{splicejam core functions}
