% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicejam-plots.R
\name{spline3d}
\alias{spline3d}
\title{Calculate a spline curve fit in 3-D}
\usage{
spline3d(x, y = NULL, z = NULL, lengthFactor = 4, verbose = FALSE,
  ...)
}
\arguments{
\item{x}{numeric vector or numeric \code{matrix} or \code{data.frame} with \code{colnames(x)}
containing \code{c("x","y","z")}.}

\item{y, z}{optional numeric vectors, used only when x is also a numeric
vector.}

\item{lengthFactor}{numeric multiplier used to control the number
of intermediate points during smoothing between each provided point.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\value{
numeric matrix with \code{colnames(x)=c("x","y","z")}.
}
\description{
Calculate a spline curve fit in 3-D
}
\details{
This function takes 3-dimensional points as input, and fits
a 3-D spline curve to connect the points. It essentially calls
\code{stats::spline()} on one dimension at a time, relative to the
row rank.
}
\seealso{
Other jam spatial functions: \code{\link{bgaPlotly3d}},
  \code{\link{dfWide2segments}}
}
\concept{jam spatial functions}
