% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gg.R
\name{stackJunctions}
\alias{stackJunctions}
\title{Stack the y-axis position of junctions}
\usage{
stackJunctions(gr, scoreColname = "score", scoreFactor = 1,
  baseline = NULL, ...)
}
\description{
Stack the y-axis position of junctions
}
