% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-data.R
\docType{data}
\name{test_cov_gr}
\alias{test_cov_gr}
\title{Sample exon coverage data GRanges}
\format{An object of class \code{GRanges} of length 4.}
\usage{
test_cov_gr
}
\description{
Sample exon coverage data GRanges
}
\details{
This dataset represents exon GRanges with an additional column
with NumericList values representing RNA-seq read coverage
across these exons.
}
\examples{
suppressPackageStartupMessages(library(GenomicRanges));
suppressPackageStartupMessages(library(ggplot2));
data(test_exon_gr);

# The steps below demonstrate how to create coverage data manually
test_cov_gr <- test_exon_gr;
set.seed(123);
cov_values <- c(250, 200, 180, 250);
values(test_cov_gr)[,"sample_A"] <- NumericList(
   lapply(seq_along(test_exon_gr), function(i){
      cov_values[i] +
         round(rnorm(width(test_exon_gr)[i])*7)
   }));
test_cov_gr;

# You can plot coverage using exoncov2polygon()
exondf <- exoncov2polygon(test_cov_gr, covNames="sample_A");
gg3 <- ggplot(exondf,
      aes(x=x, y=y, group=gr, fill=gr, color=gr)) +
   ggforce::geom_shape(alpha=0.8) +
   colorjam::theme_jam() +
   colorjam::scale_fill_jam() +
   colorjam::scale_color_jam();
print(gg3);

# you can add junctions to exons in one plot
gg3 +
   ggforce::geom_diagonal_wide(data=junc_df,
      inherit.aes=FALSE,
      show.legend=FALSE,
      fill="orange",
      aes(x=x, y=y, group=id),
      alpha=0.6)
}
\keyword{datasets}
