% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-data.R
\docType{data}
\name{test_exon_gr}
\alias{test_exon_gr}
\title{Sample exon data GRanges}
\format{GRanges object where each segment represents one
exon for an arbitrary gene. It has one column of values,
\code{"gene_name"} used for Sashimi plot preparation.}
\usage{
test_exon_gr
}
\description{
Sample exon data GRanges
}
\details{
This dataset contains RNA-seq splice junction data
stored as a GRangesList.
}
\examples{
# The code below is used to create the exon test data
suppressPackageStartupMessages(library(GenomicRanges));
suppressPackageStartupMessages(library(ggplot2));

test_exon_gr <- GRanges(seqnames=rep("chr1", 4),
   ranges=IRanges::IRanges(
      start=c(100, 300, 500, 900),
        end=c(199, 399, 749, 999)),
   strand=rep("+", 4),
   gene_name=rep("TestGene1", 4));
names(test_exon_gr) <- jamba::makeNames(rep("exon", length(test_exon_gr)),
   suffix="");

# To plot a simple GRanges object
ggplot(grl2df(test_exon_gr), aes(x=x, y=y, group=id, fill=feature_type)) +
   geom_polygon() +
   colorjam::theme_jam() +
   colorjam::scale_fill_jam()
}
\seealso{
Other splicejam data: \code{\link{test_cov_gr}},
  \code{\link{test_cov_wide_gr}},
  \code{\link{test_exon_wide_gr}},
  \code{\link{test_junc_gr}},
  \code{\link{test_junc_wide_gr}}
}
\concept{splicejam data}
\keyword{datasets}
