% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambio-gg.R
\name{to_basic.GeomShape}
\alias{to_basic.GeomShape}
\title{Support plotly for GeomShape}
\usage{
to_basic.GeomShape(data, prestats_data, layout, params, p, ...)
}
\arguments{
\item{data, prestats_data, layout, params, p, ...}{arguments provided
by plotly during rendering, after stats have been applied.
Currently only \code{data} is passed to \code{ggplot2::GeomPolygon}.}
}
\description{
Support plotly for GeomShape
}
\details{
This function is a helper function intended to provide a simple
but not fully-equivalent link between \code{ggforce::geom_shape()} and
\code{plotly::ggplotly()}. This function converts data to
\code{ggplot2::geom_polygon} which is sufficient for the splicejam
package, but which does not provide the extra effects from
\code{ggforce::geom_shape()} such as rounded corners or resizing.
This function may be extended then submitted to
the \code{ggforce} package to assist plotly support.

This function is not exported, as it is only used by plotly
specifically during conversion of a ggplot object to plotly object.
}
